@@ -1,20 +1,27 @@
 import setuptools
-version = '1.8.12'
+version = '1.14.0'
 
 setuptools.setup(
     name='electrumX',
     version=version,
-    scripts=['electrumx_server', 'electrumx_rpc'],
-    python_requires='>=3.6',
-    # via environment variables, in which case I've tested with 15.0.4
-    # "x11_hash" package (1.4) is required to sync DASH network.
-    # "tribus_hash" package is required to sync Denarius network.
-    # "blake256" package is required to sync Decred network.
-    # "xevan_hash" package is required to sync Xuez network.
-    # "groestlcoin_hash" package is required to sync Groestlcoin network.
-    # "pycryptodomex" package is required to sync SmartCash network.
-    install_requires=['pycryptodomex', 'aiorpcX>=0.10.1,<0.11', 'attrs',
-                      'plyvel', 'pylru', 'aiohttp >= 2'],
+    scripts=['electrumx_server', 'electrumx_rpc', 'electrumx_compact_history'],
+    python_requires='>=3.7',
+    install_requires=['aiorpcX[ws]>=0.18.3,<0.19', 'attrs',
+                      'plyvel', 'pylru', 'aiohttp>=3.3'],
+    extras_require={
+        'rocksdb': ['python-rocksdb>=0.6.9'],
+        'uvloop': ['uvloop>=0.14'],
+        # For various coins
+        'blake256': ['blake256>=0.1.1'],
+        'crypto': ['pycryptodomex>=3.8.1'],
+        'groestl': ['groestlcoin-hash>=1.0.1'],
+        'tribus-hash': ['tribus-hash>=1.0.2'],
+        'xevan-hash': ['xevan-hash'],
+        'x11-hash': ['x11-hash>=1.4'],
+        'zny-yespower-0-5': ['zny-yespower-0-5'],
+        'bell-yespower': ['bell-yespower'],
+        'cpupower': ['cpupower'],
+    },
     packages=setuptools.find_packages(include=('electrumx*',)),
     description='ElectrumX Server',
     author='Neil Booth',
@@ -29,7 +36,7 @@ setuptools.setup(
         'Framework :: AsyncIO',
         'License :: OSI Approved :: MIT License',
         'Operating System :: Unix',
-        "Programming Language :: Python :: 3.6",
+        "Programming Language :: Python :: 3.7",
         "Topic :: Database",
         'Topic :: Internet',
     ],
