@@ -5,12 +5,16 @@
 # See the file "LICENCE" for information about the copyright
 # and warranty status of this software.
 
+'''Base class of servers'''
+
 import asyncio
 import os
+import platform
 import re
 import signal
 import sys
 import time
+from contextlib import suppress
 from functools import partial
 
 from aiorpcx import spawn
@@ -18,7 +22,7 @@ from aiorpcx import spawn
 from electrumx.lib.util import class_logger
 
 
-class ServerBase(object):
+class ServerBase:
     '''Base class server implementation.
 
     Derived classes are expected to:
@@ -42,15 +46,19 @@ class ServerBase(object):
         asyncio.set_event_loop_policy(env.loop_policy)
 
         self.logger = class_logger(__name__, self.__class__.__name__)
-        self.logger.info(f'Python version: {sys.version}')
+        version_str = ' '.join(sys.version.splitlines())
+        self.logger.info(f'Python version: {version_str}')
         self.env = env
+        self.start_time = 0
 
         # Sanity checks
         if sys.version_info < self.PYTHON_MIN_VERSION:
             mvs = '.'.join(str(part) for part in self.PYTHON_MIN_VERSION)
             raise RuntimeError('Python version >= {} is required'.format(mvs))
 
-        if os.geteuid() == 0 and not env.allow_root:
+        if platform.system() == 'Windows':
+            pass
+        elif os.geteuid() == 0 and not env.allow_root:
             raise RuntimeError('RUNNING AS ROOT IS STRONGLY DISCOURAGED!\n'
                                'You shoud create an unprivileged user account '
                                'and use that.\n'
@@ -63,7 +71,6 @@ class ServerBase(object):
 
         Setting the event also shuts down the server.
         '''
-        shutdown_event.set()
 
     def on_exception(self, loop, context):
         '''Suppress spurious messages it appears we cannot control.'''
@@ -74,7 +81,7 @@ class ServerBase(object):
             return
         loop.default_exception_handler(context)
 
-    async def _main(self, loop):
+    async def run(self):
         '''Run the server application:
 
         - record start time
@@ -84,30 +91,37 @@ class ServerBase(object):
         '''
         def on_signal(signame):
             shutdown_event.set()
-            self.logger.warning(f'received {signame} signal, '
-                                f'initiating shutdown')
+            self.logger.warning(f'received {signame} signal, initiating shutdown')
+
+        async def serve():
+            try:
+                await self.serve(shutdown_event)
+            finally:
+                shutdown_event.set()
 
         self.start_time = time.time()
-        for signame in ('SIGINT', 'SIGTERM'):
-            loop.add_signal_handler(getattr(signal, signame),
-                                    partial(on_signal, signame))
+        loop = asyncio.get_event_loop()
+        shutdown_event = asyncio.Event()
+
+        if platform.system() != 'Windows':
+            # No signals on Windows
+            for signame in ('SIGINT', 'SIGTERM'):
+                loop.add_signal_handler(getattr(signal, signame),
+                                        partial(on_signal, signame))
         loop.set_exception_handler(self.on_exception)
 
-        shutdown_event = asyncio.Event()
-        server_task = await spawn(self.serve(shutdown_event))
-        # Wait for shutdown, log on receipt of the event
-        await shutdown_event.wait()
-        self.logger.info('shutting down')
-        server_task.cancel()
-
-        # Prevent some silly logs
-        await asyncio.sleep(0.01)
-
-        self.logger.info('shutdown complete')
-
-    def run(self):
-        loop = asyncio.get_event_loop()
+        # Start serving and wait for shutdown, log receipt of the event
+        server_task = await spawn(serve, report_crash=False)
         try:
-            loop.run_until_complete(self._main(loop))
+            await shutdown_event.wait()
+        except KeyboardInterrupt:
+            self.logger.warning('received keyboard interrupt, initiating shutdown')
+
+        self.logger.info('shutting down')
+
+        server_task.cancel()
+        try:
+            with suppress(asyncio.CancelledError):
+                await server_task
         finally:
-            loop.run_until_complete(loop.shutdown_asyncgens())
+            self.logger.info('shutdown complete')
