@@ -9,18 +9,18 @@
 '''Block prefetcher and chain processor.'''
 
 
-import array
 import asyncio
-from struct import pack, unpack
 import time
-from functools import partial
 
 from aiorpcx import TaskGroup, run_in_thread
 
 import electrumx
 from electrumx.server.daemon import DaemonError
 from electrumx.lib.hash import hash_to_hex_str, HASHX_LEN
-from electrumx.lib.util import chunks, class_logger
+from electrumx.lib.script import is_unspendable_legacy, is_unspendable_genesis
+from electrumx.lib.util import (
+    chunks, class_logger, pack_le_uint32, pack_le_uint64, unpack_le_uint32
+)
 from electrumx.server.db import FlushData
 
 
@@ -57,6 +57,11 @@ class Prefetcher(object):
                     await asyncio.sleep(self.polling_delay)
             except DaemonError as e:
                 self.logger.info(f'ignoring daemon error: {e}')
+            except asyncio.CancelledError as e:
+                self.logger.info(f'cancelled; prefetcher stopping {e}')
+                raise
+            except Exception:
+                self.logger.exception(f'ignoring unexpected exception')
 
     def get_prefetched_blocks(self):
         '''Called by block processor when it is processing queued blocks.'''
@@ -98,17 +103,16 @@ class Prefetcher(object):
         daemon_height = await daemon.height()
         async with self.semaphore:
             while self.cache_size < self.min_cache_size:
+                first = self.fetched_height + 1
                 # Try and catch up all blocks but limit to room in cache.
-                # Constrain fetch count to between 0 and 500 regardless;
-                # testnet can be lumpy.
                 cache_room = max(self.min_cache_size // self.ave_size, 1)
                 count = min(daemon_height - self.fetched_height, cache_room)
-                count = min(500, count)
+                # Don't make too large a request
+                count = min(self.coin.max_fetch_blocks(first), max(count, 0))
                 if not count:
                     self.caught_up = True
                     return False
 
-                first = self.fetched_height + 1
                 hex_hashes = await daemon.block_hex_hashes(first, count)
                 if self.caught_up:
                     self.logger.info('new block height {:,d} hash {}'
@@ -165,6 +169,10 @@ class BlockProcessor(object):
         self.next_cache_check = 0
         self.touched = set()
         self.reorg_count = 0
+        self.height = -1
+        self.tip = None
+        self.tx_count = 0
+        self._caught_up_event = None
 
         # Caches of unflushed items.
         self.headers = []
@@ -209,9 +217,9 @@ class BlockProcessor(object):
             await self._maybe_flush()
             if not self.db.first_sync:
                 s = '' if len(blocks) == 1 else 's'
-                self.logger.info('processed {:,d} block{} in {:.1f}s'
-                                 .format(len(blocks), s,
-                                         time.time() - start))
+                blocks_size = sum(len(block) for block in raw_blocks) / 1_000_000
+                self.logger.info(f'processed {len(blocks):,d} block{s} size {blocks_size:.2f} MB '
+                                 f'in {time.time() - start:.1f}s')
             if self._caught_up_event.is_set():
                 await self.notifications.on_block(self.touched, self.height)
             self.touched = set()
@@ -253,7 +261,7 @@ class BlockProcessor(object):
             self.touched.discard(None)
             self.db.flush_backup(self.flush_data(), self.touched)
 
-        start, last, hashes = await self.reorg_hashes(count)
+        _start, last, hashes = await self.reorg_hashes(count)
         # Reverse and convert to hex strings.
         hashes = [hash_to_hex_str(hash) for hash in reversed(hashes)]
         for hex_hashes in chunks(hashes, 50):
@@ -378,10 +386,13 @@ class BlockProcessor(object):
         '''
         min_height = self.db.min_undo_height(self.daemon.cached_height())
         height = self.height
+        genesis_activation = self.coin.GENESIS_ACTIVATION
 
         for block in blocks:
             height += 1
-            undo_info = self.advance_txs(block.transactions)
+            is_unspendable = (is_unspendable_genesis if height >= genesis_activation
+                              else is_unspendable_legacy)
+            undo_info = self.advance_txs(block.transactions, is_unspendable)
             if height >= min_height:
                 self.undo_infos.append((undo_info, height))
                 self.db.write_raw_block(block.raw, height)
@@ -391,25 +402,26 @@ class BlockProcessor(object):
         self.headers.extend(headers)
         self.tip = self.coin.header_hash(headers[-1])
 
-    def advance_txs(self, txs):
+    def advance_txs(self, txs, is_unspendable):
         self.tx_hashes.append(b''.join(tx_hash for tx, tx_hash in txs))
 
         # Use local vars for speed in the loops
         undo_info = []
         tx_num = self.tx_count
         script_hashX = self.coin.hashX_from_script
-        s_pack = pack
         put_utxo = self.utxo_cache.__setitem__
         spend_utxo = self.spend_utxo
         undo_info_append = undo_info.append
         update_touched = self.touched.update
         hashXs_by_tx = []
         append_hashXs = hashXs_by_tx.append
+        to_le_uint32 = pack_le_uint32
+        to_le_uint64 = pack_le_uint64
 
         for tx, tx_hash in txs:
             hashXs = []
             append_hashX = hashXs.append
-            tx_numb = s_pack('<I', tx_num)
+            tx_numb = to_le_uint32(tx_num)
 
             # Spend the inputs
             for txin in tx.inputs:
@@ -421,12 +433,15 @@ class BlockProcessor(object):
 
             # Add the new UTXOs
             for idx, txout in enumerate(tx.outputs):
-                # Get the hashX.  Ignore unspendable outputs
+                # Ignore unspendable outputs
+                if is_unspendable(txout.pk_script):
+                    continue
+
+                # Get the hashX
                 hashX = script_hashX(txout.pk_script)
-                if hashX:
-                    append_hashX(hashX)
-                    put_utxo(tx_hash + s_pack('<H', idx),
-                             hashX + tx_numb + s_pack('<Q', txout.value))
+                append_hashX(hashX)
+                put_utxo(tx_hash + to_le_uint32(idx),
+                         hashX + tx_numb + to_le_uint64(txout.value))
 
             append_hashXs(hashXs)
             update_touched(hashXs)
@@ -447,6 +462,7 @@ class BlockProcessor(object):
         '''
         self.db.assert_flushed(self.flush_data())
         assert self.height >= len(raw_blocks)
+        genesis_activation = self.coin.GENESIS_ACTIVATION
 
         coin = self.coin
         for raw_block in raw_blocks:
@@ -459,13 +475,15 @@ class BlockProcessor(object):
                                          hash_to_hex_str(self.tip),
                                          self.height))
             self.tip = coin.header_prevhash(block.header)
-            self.backup_txs(block.transactions)
+            is_unspendable = (is_unspendable_genesis if self.height >= genesis_activation
+                              else is_unspendable_legacy)
+            self.backup_txs(block.transactions, is_unspendable)
             self.height -= 1
             self.db.tx_counts.pop()
 
         self.logger.info('backed up to height {:,d}'.format(self.height))
 
-    def backup_txs(self, txs):
+    def backup_txs(self, txs, is_unspendable):
         # Prevout values, in order down the block (coinbase first if present)
         # undo_info is in reverse block order
         undo_info = self.db.read_undo_info(self.height)
@@ -475,7 +493,6 @@ class BlockProcessor(object):
         n = len(undo_info)
 
         # Use local vars for speed in the loops
-        s_pack = pack
         put_utxo = self.utxo_cache.__setitem__
         spend_utxo = self.spend_utxo
         script_hashX = self.coin.hashX_from_script
@@ -486,10 +503,13 @@ class BlockProcessor(object):
             for idx, txout in enumerate(tx.outputs):
                 # Spend the TX outputs.  Be careful with unspendable
                 # outputs - we didn't save those in the first place.
+                if is_unspendable(txout.pk_script):
+                    continue
+
+                # Get the hashX
                 hashX = script_hashX(txout.pk_script)
-                if hashX:
-                    cache_value = spend_utxo(tx_hash, idx)
-                    touched.add(cache_value[:-12])
+                cache_value = spend_utxo(tx_hash, idx)
+                touched.add(cache_value[:-12])
 
             # Restore the inputs
             for txin in reversed(tx.inputs):
@@ -497,8 +517,7 @@ class BlockProcessor(object):
                     continue
                 n -= undo_entry_len
                 undo_item = undo_info[n:n + undo_entry_len]
-                put_utxo(txin.prev_hash + s_pack('<H', txin.prev_idx),
-                         undo_item)
+                put_utxo(txin.prev_hash + pack_le_uint32(txin.prev_idx), undo_item)
                 touched.add(undo_item[:-12])
 
         assert n == 0
@@ -514,10 +533,10 @@ class BlockProcessor(object):
     TX not per UTXO).  So store them in a Python dictionary with
     binary keys and values.
 
-      Key:    TX_HASH + TX_IDX           (32 + 2 = 34 bytes)
+      Key:    TX_HASH + TX_IDX           (32 + 4 = 36 bytes)
       Value:  HASHX + TX_NUM + VALUE     (11 + 4 + 8 = 23 bytes)
 
-    That's 57 bytes of raw data in-memory.  Python dictionary overhead
+    That's 59 bytes of raw data in-memory.  Python dictionary overhead
     means each entry actually uses about 205 bytes of memory.  So
     almost 5 million UTXOs can fit in 1GB of RAM.  There are
     approximately 42 million UTXOs on bitcoin mainnet at height
@@ -566,7 +585,7 @@ class BlockProcessor(object):
         corruption.
         '''
         # Fast track is it being in the cache
-        idx_packed = pack('<H', tx_idx)
+        idx_packed = pack_le_uint32(tx_idx)
         cache_value = self.utxo_cache.pop(tx_hash + idx_packed, None)
         if cache_value:
             return cache_value
@@ -583,15 +602,15 @@ class BlockProcessor(object):
             tx_num_packed = hdb_key[-4:]
 
             if len(candidates) > 1:
-                tx_num, = unpack('<I', tx_num_packed)
-                hash, height = self.db.fs_tx_hash(tx_num)
+                tx_num, = unpack_le_uint32(tx_num_packed)
+                hash, _height = self.db.fs_tx_hash(tx_num)
                 if hash != tx_hash:
                     assert hash is not None  # Should always be found
                     continue
 
             # Key: b'u' + address_hashX + tx_idx + tx_num
             # Value: the UTXO value as a 64-bit unsigned integer
-            udb_key = b'u' + hashX + hdb_key[-6:]
+            udb_key = b'u' + hashX + hdb_key[-8:]
             utxo_value_packed = self.db.utxo_db.get(udb_key)
             if utxo_value_packed:
                 # Remove both entries for this UTXO
@@ -683,9 +702,10 @@ class DecredBlockProcessor(BlockProcessor):
         return start, count
 
 
-class NamecoinBlockProcessor(BlockProcessor):
-    def advance_txs(self, txs):
-        result = super().advance_txs(txs)
+class NameIndexBlockProcessor(BlockProcessor):
+
+    def advance_txs(self, txs, is_unspendable):
+        result = super().advance_txs(txs, is_unspendable)
 
         tx_num = self.tx_count - len(txs)
         script_name_hashX = self.coin.name_hashX_from_script
@@ -693,12 +713,12 @@ class NamecoinBlockProcessor(BlockProcessor):
         hashXs_by_tx = []
         append_hashXs = hashXs_by_tx.append
 
-        for tx, tx_hash in txs:
+        for tx, _tx_hash in txs:
             hashXs = []
             append_hashX = hashXs.append
 
             # Add the new UTXOs and associate them with the name script
-            for idx, txout in enumerate(tx.outputs):
+            for txout in tx.outputs:
                 # Get the hashX of the name script.  Ignore non-name scripts.
                 hashX = script_name_hashX(txout.pk_script)
                 if hashX:
@@ -711,3 +731,103 @@ class NamecoinBlockProcessor(BlockProcessor):
         self.db.history.add_unflushed(hashXs_by_tx, self.tx_count - len(txs))
 
         return result
+
+
+class LTORBlockProcessor(BlockProcessor):
+
+    def advance_txs(self, txs, is_unspendable):
+        self.tx_hashes.append(b''.join(tx_hash for tx, tx_hash in txs))
+
+        # Use local vars for speed in the loops
+        undo_info = []
+        tx_num = self.tx_count
+        script_hashX = self.coin.hashX_from_script
+        put_utxo = self.utxo_cache.__setitem__
+        spend_utxo = self.spend_utxo
+        undo_info_append = undo_info.append
+        update_touched = self.touched.update
+        to_le_uint32 = pack_le_uint32
+        to_le_uint64 = pack_le_uint64
+
+        hashXs_by_tx = [set() for _ in txs]
+
+        # Add the new UTXOs
+        for (tx, tx_hash), hashXs in zip(txs, hashXs_by_tx):
+            add_hashXs = hashXs.add
+            tx_numb = to_le_uint32(tx_num)
+
+            for idx, txout in enumerate(tx.outputs):
+                # Ignore unspendable outputs
+                if is_unspendable(txout.pk_script):
+                    continue
+
+                # Get the hashX
+                hashX = script_hashX(txout.pk_script)
+                add_hashXs(hashX)
+                put_utxo(tx_hash + to_le_uint32(idx),
+                         hashX + tx_numb + to_le_uint64(txout.value))
+            tx_num += 1
+
+        # Spend the inputs
+        # A separate for-loop here allows any tx ordering in block.
+        for (tx, tx_hash), hashXs in zip(txs, hashXs_by_tx):
+            add_hashXs = hashXs.add
+            for txin in tx.inputs:
+                if txin.is_generation():
+                    continue
+                cache_value = spend_utxo(txin.prev_hash, txin.prev_idx)
+                undo_info_append(cache_value)
+                add_hashXs(cache_value[:-12])
+
+        # Update touched set for notifications
+        for hashXs in hashXs_by_tx:
+            update_touched(hashXs)
+
+        self.db.history.add_unflushed(hashXs_by_tx, self.tx_count)
+
+        self.tx_count = tx_num
+        self.db.tx_counts.append(tx_num)
+
+        return undo_info
+
+    def backup_txs(self, txs, is_unspendable):
+        undo_info = self.db.read_undo_info(self.height)
+        if undo_info is None:
+            raise ChainError('no undo information found for height {:,d}'
+                             .format(self.height))
+
+        # Use local vars for speed in the loops
+        put_utxo = self.utxo_cache.__setitem__
+        spend_utxo = self.spend_utxo
+        script_hashX = self.coin.hashX_from_script
+        add_touched = self.touched.add
+        undo_entry_len = 12 + HASHX_LEN
+
+        # Restore coins that had been spent
+        # (may include coins made then spent in this block)
+        n = 0
+        for tx, tx_hash in txs:
+            for txin in tx.inputs:
+                if txin.is_generation():
+                    continue
+                undo_item = undo_info[n:n + undo_entry_len]
+                put_utxo(txin.prev_hash + pack_le_uint32(txin.prev_idx), undo_item)
+                add_touched(undo_item[:-12])
+                n += undo_entry_len
+
+        assert n == len(undo_info)
+
+        # Remove tx outputs made in this block, by spending them.
+        for tx, tx_hash in txs:
+            for idx, txout in enumerate(tx.outputs):
+                # Spend the TX outputs.  Be careful with unspendable
+                # outputs - we didn't save those in the first place.
+                if is_unspendable(txout.pk_script):
+                    continue
+
+                # Get the hashX
+                hashX = script_hashX(txout.script)
+                cache_value = spend_utxo(tx_hash, idx)
+                add_touched(cache_value[:-12])
+
+        self.tx_count -= len(txs)
