@@ -15,7 +15,7 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
-VERSION="ElectrumX 1.8.12"
+VERSION="ElectrumX 1.14.0"
 
 # -- Project information -----------------------------------------------------
 
