@@ -36,21 +36,22 @@ import struct
 from decimal import Decimal
 from hashlib import sha256
 from functools import partial
-import base64
 
 import electrumx.lib.util as util
 from electrumx.lib.hash import Base58, hash160, double_sha256, hash_to_hex_str
 from electrumx.lib.hash import HASHX_LEN, hex_str_to_hash
-from electrumx.lib.script import ScriptPubKey, OpCodes
+from electrumx.lib.script import (_match_ops, Script, ScriptError,
+                                  ScriptPubKey, OpCodes)
 import electrumx.lib.tx as lib_tx
+import electrumx.lib.tx_dash as lib_tx_dash
+import electrumx.lib.tx_axe as lib_tx_axe
 import electrumx.server.block_processor as block_proc
 import electrumx.server.daemon as daemon
 from electrumx.server.session import (ElectrumX, DashElectrumX,
-                                      SmartCashElectrumX)
+                                      SmartCashElectrumX, AuxPoWElectrumX)
 
 
 Block = namedtuple("Block", "raw header transactions")
-OP_RETURN = OpCodes.OP_RETURN
 
 
 class CoinError(Exception):
@@ -68,6 +69,7 @@ class Coin(object):
     BASIC_HEADER_SIZE = 80
     STATIC_BLOCK_HEADERS = True
     SESSIONCLS = ElectrumX
+    DEFAULT_MAX_SEND = 1000000
     DESERIALIZER = lib_tx.Deserializer
     DAEMON = daemon.Daemon
     BLOCK_PROCESSOR = block_proc.BlockProcessor
@@ -75,13 +77,21 @@ class Coin(object):
                      'bits', 'nonce')
     HEADER_UNPACK = struct.Struct('< I 32s 32s I I I').unpack_from
     MEMPOOL_HISTOGRAM_REFRESH_SECS = 500
+    P2PKH_VERBYTE = bytes.fromhex("00")
+    P2SH_VERBYTES = [bytes.fromhex("05")]
     XPUB_VERBYTES = bytes('????', 'utf-8')
     XPRV_VERBYTES = bytes('????', 'utf-8')
+    WIF_BYTE = bytes.fromhex("80")
     ENCODE_CHECK = Base58.encode_check
     DECODE_CHECK = Base58.decode_check
+    GENESIS_HASH = ('000000000019d6689c085ae165831e93'
+                    '4ff763ae46a2a6c172b3f1b60a8ce26f')
+    GENESIS_ACTIVATION = 100_000_000
     # Peer discovery
     PEER_DEFAULT_PORTS = {'t': '50001', 's': '50002'}
     PEERS = []
+    CRASH_CLIENT_VER = None
+    BLACKLIST_URL = None
 
     @classmethod
     def lookup_coin_class(cls, name, net):
@@ -116,6 +126,12 @@ class Coin(object):
         return url + '/'
 
     @classmethod
+    def max_fetch_blocks(cls, height):
+        if height < 130000:
+            return 1000
+        return 100
+
+    @classmethod
     def genesis_block(cls, block):
         '''Check the Genesis block is the right one for this coin.
 
@@ -131,11 +147,7 @@ class Coin(object):
 
     @classmethod
     def hashX_from_script(cls, script):
-        '''Returns a hashX from a script, or None if the script is provably
-        unspendable so the output can be dropped.
-        '''
-        if script and script[0] == OP_RETURN:
-            return None
+        '''Returns a hashX from a script.'''
         return sha256(script).digest()[:HASHX_LEN]
 
     @staticmethod
@@ -255,19 +267,19 @@ class Coin(object):
         return Decimal(value) / cls.VALUE_PER_COIN
 
     @classmethod
-    def electrum_header(cls, header, height):
-        h = dict(zip(cls.HEADER_VALUES, cls.HEADER_UNPACK(header)))
-        # Add the height that is not present in the header itself
-        h['block_height'] = height
-        # Convert bytes to str
-        h['prev_block_hash'] = hash_to_hex_str(h['prev_block_hash'])
-        h['merkle_root'] = hash_to_hex_str(h['merkle_root'])
-        return h
+    def warn_old_client_on_tx_broadcast(cls, _client_ver):
+        return False
 
 
 class AuxPowMixin(object):
     STATIC_BLOCK_HEADERS = False
     DESERIALIZER = lib_tx.DeserializerAuxPow
+    SESSIONCLS = AuxPoWElectrumX
+    TRUNCATED_HEADER_SIZE = 80
+    # AuxPoW headers are significantly larger, so the DEFAULT_MAX_SEND from
+    # Bitcoin is insufficient.  In Namecoin mainnet, 5 MB wasn't enough to
+    # sync, while 10 MB worked fine.
+    DEFAULT_MAX_SEND = 10000000
 
     @classmethod
     def header_hash(cls, header):
@@ -278,7 +290,7 @@ class AuxPowMixin(object):
     def block_header(cls, block, height):
         '''Return the AuxPow block header bytes'''
         deserializer = cls.DESERIALIZER(block)
-        return deserializer.read_header(height, cls.BASIC_HEADER_SIZE)
+        return deserializer.read_header(cls.BASIC_HEADER_SIZE)
 
 
 class EquihashMixin(object):
@@ -290,22 +302,10 @@ class EquihashMixin(object):
     HEADER_UNPACK = struct.Struct('< I 32s 32s 32s I I 32s').unpack_from
 
     @classmethod
-    def electrum_header(cls, header, height):
-        h = dict(zip(cls.HEADER_VALUES, cls.HEADER_UNPACK(header)))
-        # Add the height that is not present in the header itself
-        h['block_height'] = height
-        # Convert bytes to str
-        h['prev_block_hash'] = hash_to_hex_str(h['prev_block_hash'])
-        h['merkle_root'] = hash_to_hex_str(h['merkle_root'])
-        h['reserved'] = hash_to_hex_str(h['reserved'])
-        h['nonce'] = hash_to_hex_str(h['nonce'])
-        return h
-
-    @classmethod
     def block_header(cls, block, height):
         '''Return the block header bytes'''
         deserializer = cls.DESERIALIZER(block)
-        return deserializer.read_header(height, cls.BASIC_HEADER_SIZE)
+        return deserializer.read_header(cls.BASIC_HEADER_SIZE)
 
 
 class ScryptMixin(object):
@@ -341,12 +341,179 @@ class BitcoinMixin(object):
     NET = "mainnet"
     XPUB_VERBYTES = bytes.fromhex("0488b21e")
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
-    P2PKH_VERBYTE = bytes.fromhex("00")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
-    WIF_BYTE = bytes.fromhex("80")
-    GENESIS_HASH = ('000000000019d6689c085ae165831e93'
-                    '4ff763ae46a2a6c172b3f1b60a8ce26f')
     RPC_PORT = 8332
+
+
+class NameMixin(object):
+    DATA_PUSH_MULTIPLE = -2
+
+    @classmethod
+    def interpret_name_prefix(cls, script, possible_ops):
+        """Interprets a potential name prefix
+
+        Checks if the given script has a name prefix.  If it has, the
+        name prefix is split off the actual address script, and its parsed
+        fields (e.g. the name) returned.
+
+        possible_ops must be an array of arrays, defining the structures
+        of name prefixes to look out for.  Each array can consist of
+        actual opcodes, -1 for ignored data placeholders, -2 for
+        multiple ignored data placeholders and strings for named placeholders.
+        Whenever a data push matches a named placeholder,
+        the corresponding value is put into a dictionary the placeholder name
+        as key, and the dictionary of matches is returned."""
+
+        try:
+            ops = Script.get_ops(script)
+        except ScriptError:
+            return None, script
+
+        name_op_count = None
+        for pops in possible_ops:
+            # Start by translating named placeholders to -1 values, and
+            # keeping track of which op they corresponded to.
+            template = []
+            named_index = {}
+
+            n = len(pops)
+            offset = 0
+            for i, op in enumerate(pops):
+                if op == cls.DATA_PUSH_MULTIPLE:
+                    # Emercoin stores value in multiple placeholders
+                    # Script structure: https://git.io/fjuRu
+                    added, template = cls._add_data_placeholders_to_template(ops[i:], template)
+                    offset += added - 1  # subtract the "DATA_PUSH_MULTIPLE" opcode
+                elif type(op) == str:
+                    template.append(-1)
+                    named_index[op] = i + offset
+                else:
+                    template.append(op)
+            n += offset
+
+            if not _match_ops(ops[:n], template):
+                continue
+
+            name_op_count = n
+            named_values = {key: ops[named_index[key]] for key in named_index}
+            break
+
+        if name_op_count is None:
+            return None, script
+
+        name_end_pos = cls.find_end_position_of_name(script, name_op_count)
+
+        address_script = script[name_end_pos:]
+        return named_values, address_script
+
+    @classmethod
+    def _add_data_placeholders_to_template(cls, opcodes, template):
+        num_dp = cls._read_data_placeholders_count(opcodes)
+        num_2drop = num_dp // 2
+        num_drop = num_dp % 2
+
+        two_drops = [OpCodes.OP_2DROP for _ in range(num_2drop)]
+        one_drops = [OpCodes.OP_DROP for _ in range(num_drop)]
+
+        elements_added = num_dp + num_2drop + num_drop
+        placeholders = [-1 for _ in range(num_dp)]
+        drops = two_drops + one_drops
+
+        return elements_added, template + placeholders + drops
+
+    @classmethod
+    def _read_data_placeholders_count(cls, opcodes):
+        data_placeholders = 0
+
+        for opcode in opcodes:
+            if type(opcode) == tuple:
+                data_placeholders += 1
+            else:
+                break
+
+        return data_placeholders
+
+    @staticmethod
+    def find_end_position_of_name(script, length):
+        """Finds the end position of the name data
+
+        Given the number of opcodes in the name prefix (length), returns the
+        index into the byte array of where the name prefix ends."""
+        n = 0
+        for _i in range(length):
+            # Content of this loop is copied from Script.get_ops's loop
+            op = script[n]
+            n += 1
+
+            if op <= OpCodes.OP_PUSHDATA4:
+                # Raw bytes follow
+                if op < OpCodes.OP_PUSHDATA1:
+                    dlen = op
+                elif op == OpCodes.OP_PUSHDATA1:
+                    dlen = script[n]
+                    n += 1
+                elif op == OpCodes.OP_PUSHDATA2:
+                    dlen, = struct.unpack('<H', script[n: n + 2])
+                    n += 2
+                else:
+                    dlen, = struct.unpack('<I', script[n: n + 4])
+                    n += 4
+                if n + dlen > len(script):
+                    raise IndexError
+                n += dlen
+
+        return n
+
+
+class NameIndexMixin(NameMixin):
+    """Shared definitions for coins that have a name index
+
+    This class defines common functions and logic for coins that have
+    a name index in addition to the index by address / script."""
+
+    BLOCK_PROCESSOR = block_proc.NameIndexBlockProcessor
+
+    @classmethod
+    def build_name_index_script(cls, name):
+        """Returns the script by which names are indexed"""
+
+        from electrumx.lib.script import Script
+
+        res = bytearray()
+        res.append(cls.OP_NAME_UPDATE)
+        res.extend(Script.push_data(name))
+        res.extend(Script.push_data(bytes([])))
+        res.append(OpCodes.OP_2DROP)
+        res.append(OpCodes.OP_DROP)
+        res.append(OpCodes.OP_RETURN)
+
+        return bytes(res)
+
+    @classmethod
+    def split_name_script(cls, script):
+        named_values, address_script = cls.interpret_name_prefix(script, cls.NAME_OPERATIONS)
+        if named_values is None or "name" not in named_values:
+            return None, address_script
+
+        name_index_script = cls.build_name_index_script(named_values["name"][1])
+        return name_index_script, address_script
+
+    @classmethod
+    def hashX_from_script(cls, script):
+        _, address_script = cls.split_name_script(script)
+        return super().hashX_from_script(address_script)
+
+    @classmethod
+    def address_from_script(cls, script):
+        _, address_script = cls.split_name_script(script)
+        return super().address_from_script(address_script)
+
+    @classmethod
+    def name_hashX_from_script(cls, script):
+        name_index_script, _ = cls.split_name_script(script)
+        if name_index_script is None:
+            return None
+
+        return super().hashX_from_script(name_index_script)
 
 
 class HOdlcoin(Coin):
@@ -355,7 +522,6 @@ class HOdlcoin(Coin):
     NET = "mainnet"
     BASIC_HEADER_SIZE = 88
     P2PKH_VERBYTE = bytes.fromhex("28")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
     WIF_BYTE = bytes.fromhex("a8")
     GENESIS_HASH = ('008872e5582924544e5c707ee4b839bb'
                     '82c28a9e94e917c94b40538d5658c04b')
@@ -365,21 +531,45 @@ class HOdlcoin(Coin):
     TX_PER_BLOCK = 5
 
 
-class BitcoinCash(BitcoinMixin, Coin):
-    NAME = "BitcoinCash"
-    SHORTNAME = "BCH"
-    TX_COUNT = 246362688
-    TX_COUNT_HEIGHT = 511484
+class BitcoinSV(BitcoinMixin, Coin):
+    NAME = "BitcoinSV"
+    SHORTNAME = "BSV"
+    TX_COUNT = 267318795
+    TX_COUNT_HEIGHT = 557037
     TX_PER_BLOCK = 400
     PEERS = [
-        'electroncash.cascharia.com s50002',
-        'bch.electrumx.cash s t',
-        'bccarihace4jdcnt.onion t52001 s52002',
-        'abc1.hsmiths.com t60001 s60002',
-        'electrumx-cash.1209k.com s t',
-        'electroncash.dk s t',
-        'electrum.imaginary.cash s t',
+        'electrumx.bitcoinsv.io s',
+        'satoshi.vision.cash s',
+        'sv.usebsv.com s t',
+        'sv.jochen-hoenicke.de s t',
+        'sv.satoshi.io s t',
     ]
+    GENESIS_ACTIVATION = 620_538
+
+
+class BitcoinCash(BitcoinMixin, Coin):
+    NAME = "BitcoinCashABC"   # Some releases later remove the ABC suffix
+    SHORTNAME = "BCH"
+    TX_COUNT = 265479628
+    TX_COUNT_HEIGHT = 556592
+    TX_PER_BLOCK = 400
+    PEERS = [
+        'bch.imaginary.cash s t',
+        'electroncash.dk s t',
+        'wallet.satoshiscoffeehouse.com s t',
+    ]
+    BLOCK_PROCESSOR = block_proc.LTORBlockProcessor
+
+    @classmethod
+    def warn_old_client_on_tx_broadcast(cls, client_ver):
+        if client_ver < (3, 3, 4):
+            return ('<br/><br/>'
+                    'Your transaction was successfully broadcast.<br/><br/>'
+                    'However, you are using a VULNERABLE version of Electron Cash.<br/>'
+                    'Download the latest version from this web site ONLY:<br/>'
+                    'https://electroncash.org/'
+                    '<br/><br/>')
+        return False
 
 
 class BitcoinSegwit(BitcoinMixin, Coin):
@@ -389,19 +579,36 @@ class BitcoinSegwit(BitcoinMixin, Coin):
     TX_COUNT = 318337769
     TX_COUNT_HEIGHT = 524213
     TX_PER_BLOCK = 1400
+    CRASH_CLIENT_VER = (3, 2, 3)
+    BLACKLIST_URL = 'https://electrum.org/blacklist.json'
     PEERS = [
-        'btc.smsys.me s995',
         'E-X.not.fyi s t',
-        'elec.luggs.co s443',
         'electrum.vom-stausee.de s t',
         'electrum.hsmiths.com s t',
         'helicarrier.bauerj.eu s t',
         'hsmiths4fyqlw5xw.onion s t',
-        'luggscoqbymhvnkp.onion t80',
         'ozahtqwp25chjdjd.onion s t',
-        'node.arihanc.com s t',
-        'arihancckjge66iv.onion s t',
+        'electrum.hodlister.co s',
+        'electrum3.hodlister.co s',
+        'btc.usebsv.com s50006',
+        'fortress.qtornado.com s443 t',
+        'ecdsa.net s110 t',
+        'e2.keff.org s t',
+        'currentlane.lovebitco.in s t',
+        'electrum.jochen-hoenicke.de s50005 t50003',
+        'vps5.hsmiths.com s',
     ]
+
+    @classmethod
+    def warn_old_client_on_tx_broadcast(cls, client_ver):
+        if client_ver < (3, 3, 3):
+            return ('<br/><br/>'
+                    'Your transaction was successfully broadcast.<br/><br/>'
+                    'However, you are using a VULNERABLE version of Electrum.<br/>'
+                    'Download the new version from the usual place:<br/>'
+                    'https://electrum.org/'
+                    '<br/><br/>')
+        return False
 
 
 class BitcoinGold(EquihashMixin, BitcoinMixin, Coin):
@@ -416,12 +623,10 @@ class BitcoinGold(EquihashMixin, BitcoinMixin, Coin):
     TX_COUNT_HEIGHT = 499923
     TX_PER_BLOCK = 50
     REORG_LIMIT = 1000
-    RPC_PORT = 8338
+    RPC_PORT = 8332
     PEERS = [
         'electrumx-eu.bitcoingold.org s50002 t50001',
-        'electrumx-us.bitcoingold.org s50002 t50001',
-        'electrumx-eu.btcgpu.org s50002 t50001',
-        'electrumx-us.btcgpu.org s50002 t50001'
+        'electrumx-us.bitcoingold.org s50002 t50001'
     ]
 
     @classmethod
@@ -432,13 +637,6 @@ class BitcoinGold(EquihashMixin, BitcoinMixin, Coin):
             return double_sha256(header)
         else:
             return double_sha256(header[:68] + header[100:112])
-
-    @classmethod
-    def electrum_header(cls, header, height):
-        h = super().electrum_header(header, height)
-        h['reserved'] = hash_to_hex_str(header[72:100])
-        h['solution'] = hash_to_hex_str(header[140:])
-        return h
 
 
 class BitcoinGoldTestnet(BitcoinGold):
@@ -452,16 +650,13 @@ class BitcoinGoldTestnet(BitcoinGold):
     TX_COUNT = 0
     TX_COUNT_HEIGHT = 1
     NET = 'testnet'
-    RPC_PORT = 18338
+    RPC_PORT = 18332
     GENESIS_HASH = ('00000000e0781ebe24b91eedc293adfe'
                     'a2f557b53ec379e78959de3853e6f9f6')
     PEERS = [
         'test-node1.bitcoingold.org s50002',
         'test-node2.bitcoingold.org s50002',
-        'test-node3.bitcoingold.org s50002',
-        'test-node1.btcgpu.org s50002',
-        'test-node2.btcgpu.org s50002',
-        'test-node3.btcgpu.org s50002'
+        'test-node3.bitcoingold.org s50002'
     ]
 
 
@@ -482,7 +677,20 @@ class BitcoinGoldRegtest(BitcoinGold):
     PEERS = []
 
 
-class Emercoin(Coin):
+class BitcoinDiamond(BitcoinSegwit, Coin):
+    NAME = "BitcoinDiamond"
+    SHORTNAME = "BCD"
+    TX_VERSION = 12
+    TX_COUNT = 274277819
+    TX_COUNT_HEIGHT = 498678
+    TX_PER_BLOCK = 50
+    REORG_LIMIT = 1000
+    PEERS = []
+    VALUE_PER_COIN = 10000000
+    DESERIALIZER = lib_tx.DeserializerBitcoinDiamondSegWit
+
+
+class Emercoin(NameMixin, Coin):
     NAME = "Emercoin"
     SHORTNAME = "EMC"
     NET = "mainnet"
@@ -490,7 +698,6 @@ class Emercoin(Coin):
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
     P2PKH_VERBYTE = bytes.fromhex("21")
     P2SH_VERBYTES = [bytes.fromhex("5c")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('00000000bcccd459d036a588d1008fce'
                     '8da3754b205736f32ddfd35350e84c2d')
     TX_COUNT = 217380620
@@ -499,9 +706,27 @@ class Emercoin(Coin):
     VALUE_PER_COIN = 1000000
     RPC_PORT = 6662
 
-    DESERIALIZER = lib_tx.DeserializerTxTimeAuxPow
+    DESERIALIZER = lib_tx.DeserializerEmercoin
 
     PEERS = []
+
+    # Name opcodes
+    OP_NAME_NEW = OpCodes.OP_1
+    OP_NAME_UPDATE = OpCodes.OP_2
+    OP_NAME_DELETE = OpCodes.OP_3
+
+    # Valid name prefixes.
+    NAME_NEW_OPS = [OP_NAME_NEW, OpCodes.OP_DROP, "name", "days",
+                    OpCodes.OP_2DROP, NameMixin.DATA_PUSH_MULTIPLE]
+    NAME_UPDATE_OPS = [OP_NAME_UPDATE, OpCodes.OP_DROP, "name", "days",
+                       OpCodes.OP_2DROP, NameMixin.DATA_PUSH_MULTIPLE]
+    NAME_DELETE_OPS = [OP_NAME_DELETE, OpCodes.OP_DROP, "name",
+                       OpCodes.OP_DROP]
+    NAME_OPERATIONS = [
+        NAME_NEW_OPS,
+        NAME_UPDATE_OPS,
+        NAME_DELETE_OPS,
+    ]
 
     @classmethod
     def block_header(cls, block, height):
@@ -509,13 +734,19 @@ class Emercoin(Coin):
         deserializer = cls.DESERIALIZER(block)
 
         if deserializer.is_merged_block():
-            return deserializer.read_header(height, cls.BASIC_HEADER_SIZE)
+            return deserializer.read_header(cls.BASIC_HEADER_SIZE)
         return block[:cls.static_header_len(height)]
 
     @classmethod
     def header_hash(cls, header):
         '''Given a header return hash'''
         return double_sha256(header[:cls.BASIC_HEADER_SIZE])
+
+    @classmethod
+    def hashX_from_script(cls, script):
+        _, address_script = cls.interpret_name_prefix(script, cls.NAME_OPERATIONS)
+
+        return super().hashX_from_script(address_script)
 
 
 class BitcoinTestnetMixin(object):
@@ -536,39 +767,88 @@ class BitcoinTestnetMixin(object):
     PEER_DEFAULT_PORTS = {'t': '51001', 's': '51002'}
 
 
+class BitcoinSVTestnet(BitcoinTestnetMixin, Coin):
+    '''Bitcoin Testnet for Bitcoin SV daemons.'''
+    NAME = "BitcoinSV"
+    PEERS = [
+        'electrontest.cascharia.com t51001 s51002',
+    ]
+    GENESIS_ACTIVATION = 1_344_302
+
+
+class BitcoinSVScalingTestnet(BitcoinSVTestnet):
+    NET = "scalingtest"
+    PEERS = [
+        'stn-server.electrumsv.io t51001 s51002',
+    ]
+    TX_COUNT = 2015
+    TX_COUNT_HEIGHT = 5711
+    TX_PER_BLOCK = 5000
+    GENESIS_ACTIVATION = 14_896
+
+    @classmethod
+    def max_fetch_blocks(cls, height):
+        if height <= 10:
+            return 100
+        return 3
+
+
 class BitcoinCashTestnet(BitcoinTestnetMixin, Coin):
     '''Bitcoin Testnet for Bitcoin Cash daemons.'''
-    NAME = "BitcoinCash"
+    NAME = "BitcoinCashABC"
     PEERS = [
-        'electrum-testnet-abc.criptolayer.net s50112',
-        'bchtestnet.arihanc.com t53001 s53002',
-        'ciiattqkgzebpp6jofjbrkhvhwmgnsfoayljdcrve2p3qmkbv3duaoyd.onion '
-        't53001 s53002',
+        'bch0.kister.net t s',
+        'testnet.imaginary.cash t50001 s50002',
+        'blackie.c3-soft.com t60001 s60002',
     ]
+    BLOCK_PROCESSOR = block_proc.LTORBlockProcessor
+
+    @classmethod
+    def warn_old_client_on_tx_broadcast(cls, client_ver):
+        if client_ver < (3, 3, 4):
+            return ('<br/><br/>'
+                    'Your transaction was successfully broadcast.<br/><br/>'
+                    'However, you are using a VULNERABLE version of Electron Cash.<br/>'
+                    'Download the latest version from this web site ONLY:<br/>'
+                    'https://electroncash.org/'
+                    '<br/><br/>')
+        return False
 
 
-class BitcoinCashRegtest(BitcoinCashTestnet):
+class BitcoinSVRegtest(BitcoinSVTestnet):
     NET = "regtest"
     GENESIS_HASH = ('0f9188f13cb7b2c71f2a335e3a4fc328'
                     'bf5beb436012afca590b1a11466e2206')
     PEERS = []
     TX_COUNT = 1
     TX_COUNT_HEIGHT = 1
+    GENESIS_ACTIVATION = 10_000
 
 
 class BitcoinSegwitTestnet(BitcoinTestnetMixin, Coin):
     '''Bitcoin Testnet for Core bitcoind >= 0.13.1.'''
     NAME = "BitcoinSegwit"
     DESERIALIZER = lib_tx.DeserializerSegWit
+    CRASH_CLIENT_VER = (3, 2, 3)
     PEERS = [
-        'electrum.akinbo.org s t',
-        'he36kyperp3kbuxu.onion s t',
         'testnet.hsmiths.com t53011 s53012',
         'hsmithsxurybd7uh.onion t53011 s53012',
-        'testnetnode.arihanc.com s t',
-        'w3e2orjpiiv2qwem3dw66d7c4krink4nhttngkylglpqe5r22n6n5wid.onion s t',
         'testnet.qtornado.com s t',
+        'testnet1.bauerj.eu t50001 s50002',
+        'tn.not.fyi t55001 s55002',
+        'bitcoin.cluelessperson.com s t',
     ]
+
+    @classmethod
+    def warn_old_client_on_tx_broadcast(cls, client_ver):
+        if client_ver < (3, 3, 3):
+            return ('<br/><br/>'
+                    'Your transaction was successfully broadcast.<br/><br/>'
+                    'However, you are using a VULNERABLE version of Electrum.<br/>'
+                    'Download the new version from the usual place:<br/>'
+                    'https://electrum.org/'
+                    '<br/><br/>')
+        return False
 
 
 class BitcoinSegwitRegtest(BitcoinSegwitTestnet):
@@ -613,7 +893,7 @@ class Litecoin(Coin):
     RPC_PORT = 9332
     REORG_LIMIT = 800
     PEERS = [
-        'elec.luggs.co s444',
+        'ex.lug.gs s444',
         'electrum-ltc.bysh.me s t',
         'electrum-ltc.ddns.net s t',
         'electrum-ltc.wilv.in s t',
@@ -643,6 +923,26 @@ class LitecoinTestnet(Litecoin):
         'electrum-ltc.bysh.me s t',
         'electrum.ltc.xurious.com s t',
     ]
+
+
+class LitecoinRegtest(LitecoinTestnet):
+    NET = "regtest"
+    GENESIS_HASH = ('530827f38f93b43ed12af0b3ad25a288'
+                    'dc02ed74d6d7857862df51fc56c416f9')
+    PEERS = []
+    TX_COUNT = 1
+    TX_COUNT_HEIGHT = 1
+
+
+class BitcoinCashRegtest(BitcoinTestnetMixin, Coin):
+    NAME = "BitcoinCashABC"   # Some releases later remove the ABC suffix
+    NET = "regtest"
+    PEERS = []
+    GENESIS_HASH = ('0f9188f13cb7b2c71f2a335e3a4fc328'
+                    'bf5beb436012afca590b1a11466e2206')
+    TX_COUNT = 1
+    TX_COUNT_HEIGHT = 1
+    BLOCK_PROCESSOR = block_proc.LTORBlockProcessor
 
 
 class Viacoin(AuxPowMixin, Coin):
@@ -689,8 +989,69 @@ class ViacoinTestnetSegWit(ViacoinTestnet):
     DESERIALIZER = lib_tx.DeserializerSegWit
 
 
+# Source: https://github.com/GravityCoinOfficial/GravityCoin/
+class GravityCoin(Coin):
+    NAME = "GravityCoin"
+    SHORTNAME = "GXX"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("28")
+    P2SH_VERBYTES = [bytes.fromhex("0a")]
+    WIF_BYTE = bytes.fromhex("d2")
+    GENESIS_HASH = ('322bad477efb4b33fa4b1f0b2861eaf543c61068da9898a95062fdb02ada486f')
+    TX_COUNT = 446050
+    TX_COUNT_HEIGHT = 547346
+    TX_PER_BLOCK = 2
+    PEER_DEFAULT_PORTS = {'t': '50001', 's': '50002'}
+    RPC_PORT = 29200
+    REORG_LIMIT = 5000
+    PEERS = []
+
+
+# Source: https://github.com/BitcoinZeroOfficial/bitcoinzero
+class Bitcoinzero(Coin):
+    NAME = "Bitcoinzero"
+    SHORTNAME = "BZX"
+    TX_COUNT = 43798
+    TX_COUNT_HEIGHT = 44
+    TX_PER_BLOCK = 576
+    NET = "mainnet"
+    GENESIS_HASH = '322bad477efb4b33fa4b1f0b2861eaf543c61068da9898a95062fdb02ada486f'
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("4b")
+    P2SH_VERBYTES = [bytes.fromhex("22")]
+    WIF_BYTE = bytes.fromhex("d2")
+    RPC_PORT = 29202
+    REORG_LIMIT = 5000
+    PEERS = []
+
+
+class Unitus(Coin):
+    NAME = "Unitus"
+    SHORTNAME = "UIS"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    P2PKH_VERBYTE = bytes.fromhex("44")
+    P2SH_VERBYTES = [bytes.fromhex("0A")]
+    WIF_BYTE = bytes.fromhex("84")
+    GENESIS_HASH = ('d8a2b2439d013a59f3bfc626a33487a3'
+                    'd7d27e42a3c9e0b81af814cd8e592f31')
+    DESERIALIZER = lib_tx.DeserializerSegWit
+    TX_COUNT = 3484561
+    TX_COUNT_HEIGHT = 1697605
+    TX_PER_BLOCK = 3
+    RPC_PORT = 50604
+    REORG_LIMIT = 2000
+    PEERS = [
+        'electrumx.unituscurrency.com s t',
+    ]
+
+
 # Source: namecoin.org
-class Namecoin(AuxPowMixin, Coin):
+class Namecoin(NameIndexMixin, AuxPowMixin, Coin):
     NAME = "Namecoin"
     SHORTNAME = "NMC"
     NET = "mainnet"
@@ -701,115 +1062,36 @@ class Namecoin(AuxPowMixin, Coin):
     WIF_BYTE = bytes.fromhex("e4")
     GENESIS_HASH = ('000000000062b72c5e2ceb45fbc8587e'
                     '807c155b0da735e6483dfba2f0a9c770')
+    DESERIALIZER = lib_tx.DeserializerAuxPowSegWit
     TX_COUNT = 4415768
     TX_COUNT_HEIGHT = 329065
     TX_PER_BLOCK = 10
+    RPC_PORT = 8336
     PEERS = [
-        'elec.luggs.co s446',
+        'electrum-nmc.le-space.de s50002',
+        'ex.lug.gs s446',
+        'luggscoqbymhvnkp.onion t82',
+        'nmc.bitcoins.sk s50002',
+        'ulrichard.ch s50006 t50005',
     ]
-    BLOCK_PROCESSOR = block_proc.NamecoinBlockProcessor
+    BLOCK_PROCESSOR = block_proc.NameIndexBlockProcessor
 
-    @classmethod
-    def split_name_script(cls, script):
-        from electrumx.lib.script import _match_ops, Script, ScriptError
+    # Name opcodes
+    OP_NAME_NEW = OpCodes.OP_1
+    OP_NAME_FIRSTUPDATE = OpCodes.OP_2
+    OP_NAME_UPDATE = OpCodes.OP_3
 
-        try:
-            ops = Script.get_ops(script)
-        except ScriptError:
-            return None, script
-
-        match = _match_ops
-
-        # Name opcodes
-        OP_NAME_NEW = OpCodes.OP_1
-        OP_NAME_FIRSTUPDATE = OpCodes.OP_2
-        OP_NAME_UPDATE = OpCodes.OP_3
-
-        # Opcode sequences for name operations
-        NAME_NEW_OPS = [OP_NAME_NEW, -1, OpCodes.OP_2DROP]
-        NAME_FIRSTUPDATE_OPS = [OP_NAME_FIRSTUPDATE, -1, -1, -1,
-                                OpCodes.OP_2DROP, OpCodes.OP_2DROP]
-        NAME_UPDATE_OPS = [OP_NAME_UPDATE, -1, -1, OpCodes.OP_2DROP,
-                           OpCodes.OP_DROP]
-
-        name_script_op_count = None
-        name_pushdata = None
-
-        # Detect name operations; determine count of opcodes.
-        # Also extract the name field -- we might use that for something in a
-        # future version.
-        if match(ops[:len(NAME_NEW_OPS)], NAME_NEW_OPS):
-            name_script_op_count = len(NAME_NEW_OPS)
-        elif match(ops[:len(NAME_FIRSTUPDATE_OPS)], NAME_FIRSTUPDATE_OPS):
-            name_script_op_count = len(NAME_FIRSTUPDATE_OPS)
-            name_pushdata = ops[1]
-        elif match(ops[:len(NAME_UPDATE_OPS)], NAME_UPDATE_OPS):
-            name_script_op_count = len(NAME_UPDATE_OPS)
-            name_pushdata = ops[1]
-
-        if name_script_op_count is None:
-            return None, script
-
-        # Find the end position of the name data
-        n = 0
-        for i in range(name_script_op_count):
-            # Content of this loop is copied from Script.get_ops's loop
-            op = script[n]
-            n += 1
-
-            if op <= OpCodes.OP_PUSHDATA4:
-                # Raw bytes follow
-                if op < OpCodes.OP_PUSHDATA1:
-                    dlen = op
-                elif op == OpCodes.OP_PUSHDATA1:
-                    dlen = script[n]
-                    n += 1
-                elif op == OpCodes.OP_PUSHDATA2:
-                    dlen, = struct.unpack('<H', script[n: n + 2])
-                    n += 2
-                else:
-                    dlen, = struct.unpack('<I', script[n: n + 4])
-                    n += 4
-                if n + dlen > len(script):
-                    raise IndexError
-                op = (op, script[n:n + dlen])
-                n += dlen
-        # Strip the name data to yield the address script
-        address_script = script[n:]
-
-        if name_pushdata is None:
-            return None, address_script
-
-        normalized_name_op_script = bytearray()
-        normalized_name_op_script.append(OP_NAME_UPDATE)
-        normalized_name_op_script.extend(Script.push_data(name_pushdata[1]))
-        normalized_name_op_script.extend(Script.push_data(bytes([])))
-        normalized_name_op_script.append(OpCodes.OP_2DROP)
-        normalized_name_op_script.append(OpCodes.OP_DROP)
-        normalized_name_op_script.append(OpCodes.OP_RETURN)
-
-        return bytes(normalized_name_op_script), address_script
-
-    @classmethod
-    def hashX_from_script(cls, script):
-        name_op_script, address_script = cls.split_name_script(script)
-
-        return super().hashX_from_script(address_script)
-
-    @classmethod
-    def address_from_script(cls, script):
-        name_op_script, address_script = cls.split_name_script(script)
-
-        return super().address_from_script(address_script)
-
-    @classmethod
-    def name_hashX_from_script(cls, script):
-        name_op_script, address_script = cls.split_name_script(script)
-
-        if name_op_script is None:
-            return None
-
-        return super().hashX_from_script(name_op_script)
+    # Valid name prefixes.
+    NAME_NEW_OPS = [OP_NAME_NEW, -1, OpCodes.OP_2DROP]
+    NAME_FIRSTUPDATE_OPS = [OP_NAME_FIRSTUPDATE, "name", -1, -1,
+                            OpCodes.OP_2DROP, OpCodes.OP_2DROP]
+    NAME_UPDATE_OPS = [OP_NAME_UPDATE, "name", -1, OpCodes.OP_2DROP,
+                       OpCodes.OP_DROP]
+    NAME_OPERATIONS = [
+        NAME_NEW_OPS,
+        NAME_FIRSTUPDATE_OPS,
+        NAME_UPDATE_OPS,
+    ]
 
 
 class NamecoinTestnet(Namecoin):
@@ -821,6 +1103,16 @@ class NamecoinTestnet(Namecoin):
     WIF_BYTE = bytes.fromhex("ef")
     GENESIS_HASH = ('00000007199508e34a9ff81e6ec0c477'
                     'a4cccff2a4767a8eee39c11db367b008')
+
+
+class NamecoinRegtest(NamecoinTestnet):
+    NAME = "Namecoin"
+    NET = "regtest"
+    GENESIS_HASH = ('0f9188f13cb7b2c71f2a335e3a4fc328'
+                    'bf5beb436012afca590b1a11466e2206')
+    PEERS = []
+    TX_COUNT = 1
+    TX_COUNT_HEIGHT = 1
 
 
 class Dogecoin(AuxPowMixin, Coin):
@@ -862,7 +1154,6 @@ class Motion(Coin):
                     '22c374496b61c1a272519b1c39979d78')
     P2PKH_VERBYTE = bytes.fromhex("32")
     P2SH_VERBYTES = [bytes.fromhex("12")]
-    WIF_BYTE = bytes.fromhex("80")
     TX_COUNT_HEIGHT = 54353
     TX_COUNT = 92701
     TX_PER_BLOCK = 4
@@ -903,6 +1194,7 @@ class Dash(Coin):
     ]
     SESSIONCLS = DashElectrumX
     DAEMON = daemon.DashDaemon
+    DESERIALIZER = lib_tx_dash.DeserializerDash
 
     @classmethod
     def header_hash(cls, header):
@@ -937,7 +1229,6 @@ class Argentum(AuxPowMixin, Coin):
     SHORTNAME = "ARG"
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("17")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
     WIF_BYTE = bytes.fromhex("97")
     GENESIS_HASH = ('88c667bc63167685e4e4da058fffdfe8'
                     'e007e5abffd6855de52ad59df7bb0bb2')
@@ -961,8 +1252,6 @@ class DigiByte(Coin):
     SHORTNAME = "DGB"
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1E")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('7497ea1b465eb39f1c8f507bc877078f'
                     'e016d6fcb6dfad3a64c98dcc6e1e8496')
     DESERIALIZER = lib_tx.DeserializerSegWit
@@ -1014,12 +1303,6 @@ class FairCoin(Coin):
         else:
             return Block(raw_block, cls.block_header(raw_block, height), [])
 
-    @classmethod
-    def electrum_header(cls, header, height):
-        h = super().electrum_header(header, height)
-        h['payload_hash'] = hash_to_hex_str(h['payload_hash'])
-        return h
-
 
 class Zcash(EquihashMixin, Coin):
     NAME = "Zcash"
@@ -1027,7 +1310,6 @@ class Zcash(EquihashMixin, Coin):
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1CB8")
     P2SH_VERBYTES = [bytes.fromhex("1CBD")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('00040fe8ec8471911baa1db1266ea15d'
                     'd06b4a8a5c453883c000b031973dce08')
     DESERIALIZER = lib_tx.DeserializerZcash
@@ -1054,27 +1336,19 @@ class ZcashTestnet(Zcash):
 
 class SnowGem(EquihashMixin, Coin):
     NAME = "SnowGem"
-    SHORTNAME = "SNG"
+    SHORTNAME = "XSG"
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1C28")
     P2SH_VERBYTES = [bytes.fromhex("1C2D")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('00068b35729d9d2b0c294ff1fe9af009'
                     '4740524311a131de40e7f705e4c29a5b')
     DESERIALIZER = lib_tx.DeserializerZcash
-    TX_COUNT = 140698
-    TX_COUNT_HEIGHT = 102802
+    TX_COUNT = 1680878
+    TX_COUNT_HEIGHT = 627250
     TX_PER_BLOCK = 2
     RPC_PORT = 16112
     REORG_LIMIT = 800
     CHUNK_SIZE = 200
-
-    @classmethod
-    def electrum_header(cls, header, height):
-        h = super().electrum_header(header, height)
-        h['n_solution'] = base64.b64encode(lib_tx.Deserializer(
-            header, start=140)._read_varbytes()).decode('utf8')
-        return h
 
 
 class BitcoinZ(EquihashMixin, Coin):
@@ -1083,7 +1357,6 @@ class BitcoinZ(EquihashMixin, Coin):
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1CB8")
     P2SH_VERBYTES = [bytes.fromhex("1CBD")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('f499ee3d498b4298ac6a64205b8addb7'
                     'c43197e2a660229be65db8a4534d75c1')
     DESERIALIZER = lib_tx.DeserializerZcash
@@ -1100,7 +1373,6 @@ class Hush(EquihashMixin, Coin):
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1CB8")
     P2SH_VERBYTES = [bytes.fromhex("1CBD")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('0003a67bc26fe564b75daf11186d3606'
                     '52eb435a35ba3d9d3e7e5d5f8e62dc17')
     DESERIALIZER = lib_tx.DeserializerZcash
@@ -1117,7 +1389,6 @@ class ZelCash(EquihashMixin, Coin):
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1CB8")
     P2SH_VERBYTES = [bytes.fromhex("1CBD")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('00052461a5006c2e3b74ce48992a0869'
                     '5607912d5604c3eb8da25749b0900444')
     DESERIALIZER = lib_tx.DeserializerZcash
@@ -1134,7 +1405,6 @@ class Zclassic(EquihashMixin, Coin):
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1CB8")
     P2SH_VERBYTES = [bytes.fromhex("1CBD")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('0007104ccda289427919efc39dc9e4d4'
                     '99804b7bebc22df55f8b834301260602')
     DESERIALIZER = lib_tx.DeserializerZcash
@@ -1151,7 +1421,6 @@ class Koto(Coin):
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("1836")
     P2SH_VERBYTES = [bytes.fromhex("183B")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('6d424c350729ae633275d51dc3496e16'
                     'cd1b1d195c164da00f39c499a2e9959e')
     DESERIALIZER = lib_tx.DeserializerZcash
@@ -1213,7 +1482,6 @@ class Einsteinium(Coin):
     SHORTNAME = "EMC2"
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("21")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
     WIF_BYTE = bytes.fromhex("b0")
     GENESIS_HASH = ('4e56204bb7b8ac06f860ff1c845f03f9'
                     '84303b5b97eb7b42868f714611aed94b')
@@ -1258,6 +1526,37 @@ class Bitbay(ScryptMixin, Coin):
     REORG_LIMIT = 5000
 
 
+class DeepOnion(Coin):
+    NAME = "DeepOnion"
+    SHORTNAME = "ONION"
+    NET = "mainnet"
+    P2PKH_VERBYTE = bytes.fromhex("1F")
+    P2SH_VERBYTES = [bytes.fromhex("4E")]
+    WIF_BYTE = bytes.fromhex("9f")
+    GENESIS_HASH = ('000004e29458ef4f2e0abab544737b07'
+                    '344e6ff13718f7c2d12926166db07b5e')
+    DESERIALIZER = lib_tx.DeserializerTxTime
+    DAEMON = daemon.LegacyRPCDaemon
+    TX_COUNT = 1194707
+    TX_COUNT_HEIGHT = 530000
+    TX_PER_BLOCK = 2
+    RPC_PORT = 18580
+    REORG_LIMIT = 200
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    PEERS = []
+
+    @classmethod
+    def header_hash(cls, header):
+        '''
+        Given a header return the hash for DeepOnion.
+        Need to download `x13_hash` module
+        Source code: https://github.com/MaruCoinOfficial/x13-hash
+        '''
+        import x13_hash
+        return x13_hash.getPoWHash(header)
+
+
 class Peercoin(Coin):
     NAME = "Peercoin"
     SHORTNAME = "PPC"
@@ -1267,13 +1566,77 @@ class Peercoin(Coin):
     WIF_BYTE = bytes.fromhex("b7")
     GENESIS_HASH = ('0000000032fe677166d54963b62a4677'
                     'd8957e87c508eaa4fd7eb1c880cd27e3')
-    DESERIALIZER = lib_tx.DeserializerTxTime
-    DAEMON = daemon.LegacyRPCDaemon
-    TX_COUNT = 1207356
-    TX_COUNT_HEIGHT = 306425
+    DESERIALIZER = lib_tx.DeserializerTxTimeSegWit
+    DAEMON = daemon.FakeEstimateFeeDaemon
+    ESTIMATE_FEE = 0.001
+    RELAY_FEE = 0.01
+    TX_COUNT = 1691771
+    TX_COUNT_HEIGHT = 455409
     TX_PER_BLOCK = 4
     RPC_PORT = 9902
     REORG_LIMIT = 5000
+
+    PEERS = [
+        "electrum.peercoinexplorer.net s"
+    ]
+
+    VALUE_PER_COIN = 1000000
+
+
+class PeercoinTestnet(Peercoin):
+    NAME = "PeercoinTestnet"
+    SHORTNAME = "tPPC"
+    NET = "testnet"
+    P2PKH_VERBYTE = bytes.fromhex("6f")
+    P2SH_VERBYTES = [bytes.fromhex("c4")]
+    WIF_BYTE = bytes.fromhex("ef")
+    GENESIS_HASH = ('00000001f757bb737f6596503e17cd17'
+                    'b0658ce630cc727c0cca81aec47c9f06')
+    ESTIMATE_FEE = 0.001
+
+
+class Trezarcoin(Coin):
+    NAME = "Trezarcoin"
+    SHORTNAME = "TZC"
+    NET = "mainnet"
+    VALUE_PER_COIN = 1000000
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    P2PKH_VERBYTE = bytes.fromhex("42")
+    P2SH_VERBYTES = [bytes.fromhex("08")]
+    WIF_BYTE = bytes.fromhex("c2")
+    GENESIS_HASH = ('24502ba55d673d2ee9170d83dae2d1ad'
+                    'b3bfb4718e4f200db9951382cc4f6ee6')
+    DESERIALIZER = lib_tx.DeserializerTrezarcoin
+    HEADER_HASH = lib_tx.DeserializerTrezarcoin.blake2s
+    HEADER_HASH_GEN = lib_tx.DeserializerTrezarcoin.blake2s_gen
+    BASIC_HEADER_SIZE = 80
+    TX_COUNT = 742886
+    TX_COUNT_HEIGHT = 643128
+    TX_PER_BLOCK = 2
+    RPC_PORT = 17299
+    REORG_LIMIT = 2000
+    PEERS = [
+        'electrumx1.trezarcoin.com s t',
+    ]
+
+    @classmethod
+    def genesis_block(cls, block):
+        '''Check the Genesis block is the right one for this coin.
+
+        Return the block less its unspendable coinbase.
+        '''
+        header = cls.block_header(block, 0)
+        header_hex_hash = cls.HEADER_HASH_GEN(header)
+        if header_hex_hash != cls.GENESIS_HASH:
+            raise CoinError('genesis block has hash {} expected {}'
+                            .format(header_hex_hash, cls.GENESIS_HASH))
+        return header + bytes(1)
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        return cls.HEADER_HASH(header)
 
 
 class Reddcoin(Coin):
@@ -1281,7 +1644,6 @@ class Reddcoin(Coin):
     SHORTNAME = "RDD"
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("3d")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
     WIF_BYTE = bytes.fromhex("bd")
     GENESIS_HASH = ('b868e0d95a3c3c0e0dadc67ee587aaf9'
                     'dc8acbf99e3b4b3110fad4eb74c1decc')
@@ -1324,8 +1686,6 @@ class Vertcoin(Coin):
     XPUB_VERBYTES = bytes.fromhex("0488B21E")
     XPRV_VERBYTES = bytes.fromhex("0488ADE4")
     P2PKH_VERBYTE = bytes.fromhex("47")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('4d96a915f49d40b1e5c2844d1ee2dccb'
                     '90013a990ccea12c492d22110489f0c4')
     DESERIALIZER = lib_tx.DeserializerSegWit
@@ -1353,14 +1713,11 @@ class Monacoin(Coin):
     TX_PER_BLOCK = 2
     RPC_PORT = 9402
     REORG_LIMIT = 1000
+    BLACKLIST_URL = 'https://electrum-mona.org/blacklist.json'
     PEERS = [
         'electrumx.tamami-foundation.org s t',
-        'electrumx2.tamami-foundation.org s t',
-        'electrumx3.tamami-foundation.org s t',
-        'electrumx1.monacoin.nl s t',
-        'electrumx2.monacoin.nl s t',
+        'electrumx3.monacoin.nl s t',
         'electrumx1.monacoin.ninja s t',
-        'electrumx2.monacoin.ninja s t',
         'electrumx2.movsign.info s t',
         'electrum-mona.bitbank.cc s t',
         'ri7rzlmdaf4eqbza.onion s t',
@@ -1389,15 +1746,22 @@ class MonacoinTestnet(Monacoin):
     ]
 
 
+class MonacoinRegtest(MonacoinTestnet):
+    NET = "regtest"
+    GENESIS_HASH = ('7543a69d7c2fcdb29a5ebec2fc064c07'
+                    '4a35253b6f3072c8a749473aa590a29c')
+    PEERS = []
+    TX_COUNT = 1
+    TX_COUNT_HEIGHT = 1
+
+
 class Crown(AuxPowMixin, Coin):
     NAME = "Crown"
     SHORTNAME = "CRW"
     NET = "mainnet"
     XPUB_VERBYTES = bytes.fromhex("0488b21e")
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
-    P2PKH_VERBYTE = bytes.fromhex("00")
     P2SH_VERBYTES = [bytes.fromhex("1c")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('0000000085370d5e122f64f4ab19c686'
                     '14ff3df78c8d13cb814fd7e69a1dc6da')
     TX_COUNT = 13336629
@@ -1444,7 +1808,6 @@ class Neblio(ScryptMixin, Coin):
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
     P2PKH_VERBYTE = bytes.fromhex("35")
     P2SH_VERBYTES = [bytes.fromhex("70")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('7286972be4dbc1463d256049b7471c25'
                     '2e6557e222cab9be73181d359cd28bcc')
     TX_COUNT = 23675
@@ -1461,13 +1824,9 @@ class Bitzeny(Coin):
     XPUB_VERBYTES = bytes.fromhex("0488b21e")
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
     P2PKH_VERBYTE = bytes.fromhex("51")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('000009f7e55e9e3b4781e22bd87a7cfa'
                     '4acada9e4340d43ca738bf4e9fb8f5ce')
-    ESTIMATE_FEE = 0.001
-    RELAY_FEE = 0.001
-    DAEMON = daemon.FakeEstimateFeeDaemon
+    DESERIALIZER = lib_tx.DeserializerSegWit
     TX_COUNT = 1408733
     TX_COUNT_HEIGHT = 1015115
     TX_PER_BLOCK = 1
@@ -1477,8 +1836,8 @@ class Bitzeny(Coin):
     @classmethod
     def header_hash(cls, header):
         '''Given a header return the hash.'''
-        import zny_yescrypt
-        return zny_yescrypt.getPoWHash(header)
+        import zny_yespower_0_5
+        return zny_yespower_0_5.getPoWHash(header)
 
 
 class CanadaeCoin(AuxPowMixin, Coin):
@@ -1488,7 +1847,6 @@ class CanadaeCoin(AuxPowMixin, Coin):
     XPUB_VERBYTES = bytes.fromhex("0488b21e")
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
     P2PKH_VERBYTE = bytes.fromhex("1C")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
     WIF_BYTE = bytes.fromhex("9c")
     GENESIS_HASH = ('863626dadaef221e2e2f30ff3dacae44'
                     'cabdae9e0028058072181b3fb675d94a')
@@ -1572,6 +1930,17 @@ class Sibcoin(Dash):
         return x11_gost_hash.getPoWHash(header)
 
 
+class SibcoinTestnet(Sibcoin):
+    SHORTNAME = "tSIB"
+    NET = "testnet"
+    XPUB_VERBYTES = bytes.fromhex("043587cf")
+    XPRV_VERBYTES = bytes.fromhex("04358394")
+    GENESIS_HASH = ('00000617791d0e19f524387f67e558b2'
+                    'a928b670b9a3b387ae003ad7f9093017')
+
+    RPC_PORT = 11944
+
+
 class Chips(Coin):
     NAME = "Chips"
     SHORTNAME = "CHIPS"
@@ -1596,7 +1965,6 @@ class Feathercoin(Coin):
     XPUB_VERBYTES = bytes.fromhex("0488BC26")
     XPRV_VERBYTES = bytes.fromhex("0488DAEE")
     P2PKH_VERBYTE = bytes.fromhex("0E")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
     WIF_BYTE = bytes.fromhex("8E")
     GENESIS_HASH = ('12a765e31ffd4059bada1e25190f6e98'
                     'c99d9714d334efa41a195a7e7e04bfe2')
@@ -1607,7 +1975,9 @@ class Feathercoin(Coin):
     RPC_PORT = 9337
     REORG_LIMIT = 2000
     PEERS = [
-        'electrumx-ch-1.feathercoin.ch s t',
+        'electrumx-gb-1.feathercoin.network s t',
+        'electrumx-gb-2.feathercoin.network s t',
+        'electrumx-de-1.feathercoin.network s t',
     ]
 
 
@@ -1667,7 +2037,6 @@ class Bitcore(BitcoinMixin, Coin):
     SHORTNAME = "BTX"
     P2PKH_VERBYTE = bytes.fromhex("03")
     P2SH_VERBYTES = [bytes.fromhex("7D")]
-    WIF_BYTE = bytes.fromhex("80")
     DESERIALIZER = lib_tx.DeserializerSegWit
     GENESIS_HASH = ('604148281e5c4b7f2487e5d03cd60d8e'
                     '6f69411d613f6448034508cea52e9574')
@@ -1675,6 +2044,12 @@ class Bitcore(BitcoinMixin, Coin):
     TX_COUNT_HEIGHT = 126946
     TX_PER_BLOCK = 2
     RPC_PORT = 8556
+    PEERS = [
+        'ele1.bitcore.cc s t',
+        'ele2.bitcore.cc s t',
+        'ele3.bitcore.cc s t',
+        'ele4.bitcore.cc s t'
+    ]
 
 
 class GameCredits(Coin):
@@ -1682,7 +2057,6 @@ class GameCredits(Coin):
     SHORTNAME = "GAME"
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("26")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
     WIF_BYTE = bytes.fromhex("a6")
     GENESIS_HASH = ('91ec5f25ee9a0ffa1af7d4da4db9a552'
                     '228dd2dc77cdb15b738be4e1f55f30ee')
@@ -1719,9 +2093,6 @@ class BitcoinAtom(Coin):
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("17")
     P2SH_VERBYTES = [bytes.fromhex("0a")]
-    WIF_BYTE = bytes.fromhex("80")
-    GENESIS_HASH = ('000000000019d6689c085ae165831e93'
-                    '4ff763ae46a2a6c172b3f1b60a8ce26f')
     STATIC_BLOCK_HEADERS = False
     DESERIALIZER = lib_tx.DeserializerBitcoinAtom
     HEADER_SIZE_POST_FORK = 84
@@ -1799,14 +2170,6 @@ class Decred(Coin):
         else:
             return Block(raw_block, cls.block_header(raw_block, height), [])
 
-    @classmethod
-    def electrum_header(cls, header, height):
-        h = super().electrum_header(header, height)
-        h['stake_root'] = hash_to_hex_str(h['stake_root'])
-        h['final_state'] = hash_to_hex_str(h['final_state'])
-        h['extra_data'] = hash_to_hex_str(h['extra_data'])
-        return h
-
 
 class DecredTestnet(Decred):
     SHORTNAME = "tDCR"
@@ -1838,7 +2201,9 @@ class Axe(Dash):
     WIF_BYTE = bytes.fromhex("cc")
     GENESIS_HASH = ('00000c33631ca6f2f61368991ce2dc03'
                     '306b5bb50bf7cede5cfbba6db38e52e6')
+    SESSIONCLS = DashElectrumX
     DAEMON = daemon.DashDaemon
+    DESERIALIZER = lib_tx_axe.DeserializerAxe
     TX_COUNT = 18405
     TX_COUNT_HEIGHT = 30237
     TX_PER_BLOCK = 1
@@ -1892,12 +2257,52 @@ class Xuez(Coin):
         else:
             return xevan_hash.getPoWHash(header)
 
+
+# Source: https://github.com/odinblockchain/odin
+class Odin(Coin):
+    NAME = "ODIN"
+    SHORTNAME = "ODIN"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("27561872")
+    XPRV_VERBYTES = bytes.fromhex("27256746")
+    P2PKH_VERBYTE = bytes.fromhex("73")
+    P2SH_VERBYTES = [bytes.fromhex("39")]
+    WIF_BYTE = bytes.fromhex("8a")
+    GENESIS_HASH = ('31ca29566549e444cf227a0e2e067aed'
+                    '847c2acc541d3bbf9ca1ae89f4fd57d7')
+
+    TX_COUNT = 340000
+    TX_COUNT_HEIGHT = 340000
+    TX_PER_BLOCK = 2
+    RPC_PORT = 22101
+    REORG_LIMIT = 100
+
+    BASIC_HEADER_SIZE = 80
+    HDR_V4_SIZE = 112
+    HDR_V4_HEIGHT = 143447
+    HDR_V4_START_OFFSET = HDR_V4_HEIGHT * BASIC_HEADER_SIZE
+
+    SESSIONCLS = DashElectrumX
+    DAEMON = daemon.DashDaemon
+    DESERIALIZER = lib_tx.DeserializerSegWit
+
     @classmethod
-    def electrum_header(cls, header, height):
-        h = super().electrum_header(header, height)
-        if h['version'] > 1:
-            h['nAccumulatorCheckpoint'] = hash_to_hex_str(header[80:])
-        return h
+    def static_header_offset(cls, height):
+        assert cls.STATIC_BLOCK_HEADERS
+        if height >= cls.HDR_V4_HEIGHT:
+            relative_v4_offset = (height - cls.HDR_V4_HEIGHT) * cls.HDR_V4_SIZE
+            return cls.HDR_V4_START_OFFSET + relative_v4_offset
+        else:
+            return height * cls.BASIC_HEADER_SIZE
+
+    @classmethod
+    def header_hash(cls, header):
+        version, = util.unpack_le_uint32_from(header)
+        if version >= 4:
+            return super().header_hash(header)
+        else:
+            import quark_hash
+            return quark_hash.getPoWHash(header)
 
 
 class Pac(Coin):
@@ -1947,6 +2352,110 @@ class PacTestnet(Pac):
     RPC_PORT = 17111
 
 
+class Zcoin(Coin):
+    NAME = "Zcoin"
+    SHORTNAME = "XZC"
+    NET = "mainnet"
+    P2PKH_VERBYTE = bytes.fromhex("52")
+    P2SH_VERBYTES = [bytes.fromhex("07")]
+    WIF_BYTE = bytes.fromhex("d2")
+    GENESIS_HASH = ('4381deb85b1b2c9843c222944b616d99'
+                    '7516dcbd6a964e1eaf0def0830695233')
+    TX_COUNT = 667154
+    TX_COUNT_HEIGHT = 100266
+    TX_PER_BLOCK = 4000  # 2000 for 1MB block
+    IRC_PREFIX = None
+    RPC_PORT = 8888
+    REORG_LIMIT = 5000
+    PEER_DEFAULT_PORTS = {'t': '50001', 's': '50002'}
+    MTP_HEADER_EXTRA_SIZE = 100
+    MTP_HEADER_DATA_SIZE = 198864
+    MTP_HEADER_DATA_START = Coin.BASIC_HEADER_SIZE + MTP_HEADER_EXTRA_SIZE
+    MTP_HEADER_DATA_END = MTP_HEADER_DATA_START + MTP_HEADER_DATA_SIZE
+    STATIC_BLOCK_HEADERS = False
+    SESSIONCLS = DashElectrumX
+    DAEMON = daemon.ZcoinMtpDaemon
+    DESERIALIZER = lib_tx.DeserializerZcoin
+    PEERS = [
+        'electrum.polispay.com'
+    ]
+
+    @classmethod
+    def is_mtp(cls, header):
+        from electrumx.lib.util import unpack_le_uint32_from, hex_to_bytes
+        if isinstance(header, str):
+            nVersion, = unpack_le_uint32_from(hex_to_bytes(header[0:4*2]))
+        elif isinstance(header, bytes):
+            nVersion, = unpack_le_uint32_from(header[0:4])
+        else:
+            raise "Cannot handle the passed type"
+        return nVersion & 0x1000
+
+    @classmethod
+    def block_header(cls, block, height):
+        sz = cls.BASIC_HEADER_SIZE
+        if cls.is_mtp(block):
+            sz += cls.MTP_HEADER_EXTRA_SIZE
+        return block[:sz]
+
+    @classmethod
+    def header_hash(cls, header):
+        sz = cls.BASIC_HEADER_SIZE
+        if cls.is_mtp(header):
+            sz += cls.MTP_HEADER_EXTRA_SIZE
+        return double_sha256(header[:sz])
+
+
+class ZcoinTestnet(Zcoin):
+    SHORTNAME = "tXZC"
+    NET = "testnet"
+    XPUB_VERBYTES = bytes.fromhex("043587cf")
+    XPRV_VERBYTES = bytes.fromhex("04358394")
+    P2PKH_VERBYTE = bytes.fromhex("41")
+    P2SH_VERBYTES = [bytes.fromhex("b2")]
+    WIF_BYTE = bytes.fromhex("b9")
+    GENESIS_HASH = '1e3487fdb1a7d46dac3e8f3e58339c6e' \
+                   'ff54abf6aef353485f3ed64250a35e89'
+    REORG_LIMIT = 8000
+    RPC_PORT = 18888
+
+
+class GINCoin(Coin):
+    NAME = "GINCoin"
+    SHORTNAME = "GIN"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    GENESIS_HASH = ('00000cd6bde619b2c3b23ad2e384328a'
+                    '450a37fa28731debf748c3b17f91f97d')
+    P2PKH_VERBYTE = bytes.fromhex("37")
+    P2SH_VERBYTES = [bytes.fromhex("38")]
+    WIF_BYTE = bytes.fromhex("3c")
+    TX_COUNT_HEIGHT = 225000
+    TX_COUNT = 470784
+    TX_PER_BLOCK = 4
+    RPC_PORT = 10211
+    PEERS = [
+        'electrum.polispay.com'
+    ]
+    SESSIONCLS = DashElectrumX
+    DAEMON = daemon.DashDaemon
+
+    # Seems that the main lyra2z_hash python package doesn't works.
+    # Tested and working with: https://github.com/LapoLab/lyra2z-py
+    @classmethod
+    def header_hash(cls, header):
+        timestamp = util.unpack_le_uint32_from(header, 68)[0]
+        if timestamp > 1550246400:
+            import x16rt_hash
+            return x16rt_hash.getPoWHash(header)
+        elif timestamp > 1525651200:
+            import lyra2z_hash
+            return lyra2z_hash.getPoWHash(header)
+        import neoscrypt
+        return neoscrypt.getPoWHash(header)
+
+
 class Polis(Coin):
     NAME = "Polis"
     SHORTNAME = "POLIS"
@@ -1958,13 +2467,12 @@ class Polis(Coin):
     P2PKH_VERBYTE = bytes.fromhex("37")
     P2SH_VERBYTES = [bytes.fromhex("38")]
     WIF_BYTE = bytes.fromhex("3c")
-    TX_COUNT_HEIGHT = 111111
-    TX_COUNT = 256128
+    TX_COUNT_HEIGHT = 280600
+    TX_COUNT = 635415
     TX_PER_BLOCK = 4
     RPC_PORT = 24127
     PEERS = [
-        'electrum1-polis.polispay.org',
-        'electrum2-polis.polispay.org'
+        'electrum.polispay.com'
     ]
     SESSIONCLS = DashElectrumX
     DAEMON = daemon.DashDaemon
@@ -1976,24 +2484,23 @@ class Polis(Coin):
         return x11_hash.getPoWHash(header)
 
 
-class ColossusXT(Coin):
-    NAME = "ColossusXT"
-    SHORTNAME = "COLX"
+class MNPCoin(Coin):
+    NAME = "MNPCoin"
+    SHORTNAME = "MNP"
     NET = "mainnet"
-    XPUB_VERBYTES = bytes.fromhex("022D2533")
-    XPRV_VERBYTES = bytes.fromhex("0221312B")
-    GENESIS_HASH = ('a0ce8206c908357008c1b9a8ba2813af'
-                    'f0989ca7f72d62b14e652c55f02b4f5c')
-    P2PKH_VERBYTE = bytes.fromhex("1E")
-    P2SH_VERBYTES = [bytes.fromhex("0D")]
-    WIF_BYTE = bytes.fromhex("D4")
-    TX_COUNT_HEIGHT = 356500
-    TX_COUNT = 761041
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    GENESIS_HASH = ('00000924036c67d803ce606ded814312'
+                    '7e62fa2111dd3b063880a1067c69ccb1')
+    P2PKH_VERBYTE = bytes.fromhex("32")
+    P2SH_VERBYTES = [bytes.fromhex("35")]
+    WIF_BYTE = bytes.fromhex("37")
+    TX_COUNT_HEIGHT = 248000
+    TX_COUNT = 506447
     TX_PER_BLOCK = 4
-    RPC_PORT = 51473
+    RPC_PORT = 13373
     PEERS = [
-        'electrum1-colx.polispay.org',
-        'electrum2-colx.polispay.org'
+        'electrum.polispay.com'
     ]
     SESSIONCLS = DashElectrumX
     DAEMON = daemon.DashDaemon
@@ -2005,62 +2512,48 @@ class ColossusXT(Coin):
         return quark_hash.getPoWHash(header)
 
 
-class GoByte(Coin):
-    NAME = "GoByte"
-    SHORTNAME = "GBX"
+class ColossusXT(Coin):
+    NAME = "ColossusXT"
+    SHORTNAME = "COLX"
     NET = "mainnet"
     XPUB_VERBYTES = bytes.fromhex("0488B21E")
     XPRV_VERBYTES = bytes.fromhex("0488ADE4")
-    GENESIS_HASH = ('0000033b01055cf8df90b01a14734cae'
-                    '92f7039b9b0e48887b4e33a469d7bc07')
-    P2PKH_VERBYTE = bytes.fromhex("26")
-    P2SH_VERBYTES = [bytes.fromhex("0A")]
-    WIF_BYTE = bytes.fromhex("C6")
-    TX_COUNT_HEIGHT = 115890
-    TX_COUNT = 245030
+    GENESIS_HASH = ('a0ce8206c908357008c1b9a8ba2813af'
+                    'f0989ca7f72d62b14e652c55f02b4f5c')
+    P2PKH_VERBYTE = bytes.fromhex("1E")
+    P2SH_VERBYTES = [bytes.fromhex("0D")]
+    WIF_BYTE = bytes.fromhex("D4")
+    TX_COUNT_HEIGHT = 356500
+    BASIC_HEADER_SIZE = 80
+    HDR_V5_HEIGHT = 500000
+    HDR_V5_SIZE = 112
+    HDR_V5_START_OFFSET = HDR_V5_HEIGHT * BASIC_HEADER_SIZE
+    TX_COUNT = 761041
     TX_PER_BLOCK = 4
-    RPC_PORT = 12454
+    RPC_PORT = 51473
     PEERS = [
-        'electrum1-gbx.polispay.org',
-        'electrum2-gbx.polispay.org'
+        'electrum.polispay.com'
     ]
     SESSIONCLS = DashElectrumX
     DAEMON = daemon.DashDaemon
 
     @classmethod
-    def header_hash(cls, header):
-        '''Given a header return the hash.'''
-        import neoscrypt
-        return neoscrypt.getPoWHash(header)
-
-
-class Monoeci(Coin):
-    NAME = "Monoeci"
-    SHORTNAME = "XMCC"
-    NET = "mainnet"
-    XPUB_VERBYTES = bytes.fromhex("0488B21E")
-    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
-    GENESIS_HASH = ('0000005be1eb05b05fb45ae38ee9c144'
-                    '1514a65343cd146100a574de4278f1a3')
-    P2PKH_VERBYTE = bytes.fromhex("32")
-    P2SH_VERBYTES = [bytes.fromhex("49")]
-    WIF_BYTE = bytes.fromhex("4D")
-    TX_COUNT_HEIGHT = 140000
-    TX_COUNT = 140000
-    TX_PER_BLOCK = 4
-    RPC_PORT = 24156
-    PEERS = [
-        'electrum1-gbx.polispay.org',
-        'electrum2-gbx.polispay.org'
-    ]
-    SESSIONCLS = DashElectrumX
-    DAEMON = daemon.DashDaemon
+    def static_header_offset(cls, height):
+        assert cls.STATIC_BLOCK_HEADERS
+        if height >= cls.HDR_V5_HEIGHT:
+            relative_v4_offset = (height - cls.HDR_V5_HEIGHT) * cls.HDR_V5_SIZE
+            return cls.HDR_V5_START_OFFSET + relative_v4_offset
+        else:
+            return height * cls.BASIC_HEADER_SIZE
 
     @classmethod
     def header_hash(cls, header):
-        '''Given a header return the hash.'''
-        import x11_hash
-        return x11_hash.getPoWHash(header)
+        version, = util.unpack_le_uint32_from(header)
+        if version >= 5:
+            return super().header_hash(header)
+        else:
+            import quark_hash
+            return quark_hash.getPoWHash(header)
 
 
 class Minexcoin(EquihashMixin, Coin):
@@ -2068,8 +2561,6 @@ class Minexcoin(EquihashMixin, Coin):
     SHORTNAME = "MNX"
     NET = "mainnet"
     P2PKH_VERBYTE = bytes.fromhex("4b")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('490a36d9451a55ed197e34aca7414b35'
                     'd775baa4a8e896f1c577f65ce2d214cb')
     STATIC_BLOCK_HEADERS = True
@@ -2081,21 +2572,16 @@ class Minexcoin(EquihashMixin, Coin):
     RPC_PORT = 8022
     CHUNK_SIZE = 960
     PEERS = [
+        'electrumx.xpresit.net s t',
         'elex01-ams.turinex.eu s t',
         'eu.minexpool.nl s t'
     ]
 
     @classmethod
-    def electrum_header(cls, header, height):
-        h = super().electrum_header(header, height)
-        h['solution'] = hash_to_hex_str(header[cls.HEADER_SIZE_NO_SOLUTION:])
-        return h
-
-    @classmethod
     def block_header(cls, block, height):
         '''Return the block header bytes'''
         deserializer = cls.DESERIALIZER(block)
-        return deserializer.read_header(height, cls.HEADER_SIZE_NO_SOLUTION)
+        return deserializer.read_header(cls.HEADER_SIZE_NO_SOLUTION)
 
 
 class Groestlcoin(Coin):
@@ -2105,8 +2591,6 @@ class Groestlcoin(Coin):
     XPUB_VERBYTES = bytes.fromhex("0488b21e")
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
     P2PKH_VERBYTE = bytes.fromhex("24")
-    P2SH_VERBYTES = [bytes.fromhex("05")]
-    WIF_BYTE = bytes.fromhex("80")
     GENESIS_HASH = ('00000ac5927c594d49cc0bdb81759d0d'
                     'a8297eb614683d3acb62f0703b639023')
     DESERIALIZER = lib_tx.DeserializerGroestlcoin
@@ -2114,6 +2598,7 @@ class Groestlcoin(Coin):
     TX_COUNT_HEIGHT = 1601528
     TX_PER_BLOCK = 5
     RPC_PORT = 1441
+    BLACKLIST_URL = 'https://groestlcoin.org/blacklist.json'
     PEERS = [
         'electrum1.groestlcoin.org s t',
         'electrum2.groestlcoin.org s t',
@@ -2154,38 +2639,37 @@ class GroestlcoinTestnet(Groestlcoin):
 
 
 class Pivx(Coin):
-    NAME = "Pivx"
+    NAME = "PIVX"
     SHORTNAME = "PIVX"
     NET = "mainnet"
     XPUB_VERBYTES = bytes.fromhex("022D2533")
     XPRV_VERBYTES = bytes.fromhex("0221312B")
+    GENESIS_HASH = ('0000041e482b9b9691d98eefb48473405c0b8ec31b76df3797c74a78680ef818')
     P2PKH_VERBYTE = bytes.fromhex("1e")
-    P2SH_VERBYTES = [bytes.fromhex("0d")]
+    P2SH_VERBYTE = bytes.fromhex("0d")
     WIF_BYTE = bytes.fromhex("d4")
-    GENESIS_HASH = ('0000041e482b9b9691d98eefb4847340'
-                    '5c0b8ec31b76df3797c74a78680ef818')
-    BASIC_HEADER_SIZE = 80
-    HDR_V4_SIZE = 112
-    HDR_V4_HEIGHT = 863787
-    HDR_V4_START_OFFSET = HDR_V4_HEIGHT * BASIC_HEADER_SIZE
-    TX_COUNT = 2930206
-    TX_COUNT_HEIGHT = 1299212
-    TX_PER_BLOCK = 2
-    RPC_PORT = 51473
+    TX_COUNT_HEIGHT = 569399
+    TX_COUNT = 2157510
+    TX_PER_BLOCK = 1
+    STATIC_BLOCK_HEADERS = False
+    RPC_PORT = 51470
+    ZEROCOIN_HEADER = 112
+    ZEROCOIN_START_HEIGHT = 863787
+    ZEROCOIN_BLOCK_VERSION = 4
 
     @classmethod
-    def static_header_offset(cls, height):
-        assert cls.STATIC_BLOCK_HEADERS
-        if height >= cls.HDR_V4_HEIGHT:
-            relative_v4_offset = (height - cls.HDR_V4_HEIGHT) * cls.HDR_V4_SIZE
-            return cls.HDR_V4_START_OFFSET + relative_v4_offset
+    def static_header_len(cls, height):
+        '''Given a header height return its length.'''
+        if (height >= cls.ZEROCOIN_START_HEIGHT):
+            return cls.ZEROCOIN_HEADER
         else:
-            return height * cls.BASIC_HEADER_SIZE
+            return cls.BASIC_HEADER_SIZE
 
     @classmethod
     def header_hash(cls, header):
-        version, = util.unpack_le_uint32_from(header)
-        if version >= 4:
+        '''Given a header return the hash.'''
+        version, = struct.unpack('<I', header[:4])
+        if version >= cls.ZEROCOIN_BLOCK_VERSION:
             return super().header_hash(header)
         else:
             import quark_hash
@@ -2193,23 +2677,16 @@ class Pivx(Coin):
 
 
 class PivxTestnet(Pivx):
-    SHORTNAME = "tPIVX"
     NET = "testnet"
     XPUB_VERBYTES = bytes.fromhex("3a8061a0")
     XPRV_VERBYTES = bytes.fromhex("3a805837")
+    GENESIS_HASH = ('0000041e482b9b9691d98eefb48473405c0b8ec31b76df3797c74a78680ef818')
     P2PKH_VERBYTE = bytes.fromhex("8B")
-    P2SH_VERBYTES = [bytes.fromhex("13")]
+    P2SH_VERBYTE = bytes.fromhex("13")
     WIF_BYTE = bytes.fromhex("EF")
-    GENESIS_HASH = (
-        '0000041e482b9b9691d98eefb48473405c0b8ec31b76df3797c74a78680ef818')
-    BASIC_HEADER_SIZE = 80
-    HDR_V4_SIZE = 112
-    HDR_V4_HEIGHT = 863787
-    HDR_V4_START_OFFSET = HDR_V4_HEIGHT * BASIC_HEADER_SIZE
-    TX_COUNT = 2157510
-    TX_COUNT_HEIGHT = 569399
-    TX_PER_BLOCK = 2
+    TX_PER_BLOCK = 4
     RPC_PORT = 51472
+    ZEROCOIN_START_HEIGHT = 201564
 
 
 class Bitg(Coin):
@@ -2253,9 +2730,9 @@ class tBitg(Bitg):
     RPC_PORT = 19332
 
 
-class CivX(Coin):
-    NAME = "CivX"
-    SHORTNAME = "CIVX"
+class EXOS(Coin):
+    NAME = "EXOS"
+    SHORTNAME = "EXOS"
     NET = "mainnet"
     XPUB_VERBYTES = bytes.fromhex("0488b21e")
     XPRV_VERBYTES = bytes.fromhex("0488ade4")
@@ -2278,11 +2755,11 @@ class CivX(Coin):
         if version > 2:
             return double_sha256(header)
         else:
-            return hex_str_to_hash(CivX.GENESIS_HASH)
+            return hex_str_to_hash(EXOS.GENESIS_HASH)
 
 
-class CivXTestnet(CivX):
-    SHORTNAME = "tCIVX"
+class EXOSTestnet(EXOS):
+    SHORTNAME = "tEXOS"
     NET = "testnet"
     XPUB_VERBYTES = bytes.fromhex("043587cf")
     XPRV_VERBYTES = bytes.fromhex("04358394")
@@ -2300,7 +2777,7 @@ class CivXTestnet(CivX):
         if version > 2:
             return double_sha256(header)
         else:
-            return hex_str_to_hash(CivXTestnet.GENESIS_HASH)
+            return hex_str_to_hash(EXOSTestnet.GENESIS_HASH)
 
 
 class SmartCash(Coin):
@@ -2330,3 +2807,556 @@ class SmartCash(Coin):
     def header_hash(cls, header):
         '''Given a header return the hash.'''
         return cls.HEADER_HASH(header)
+
+
+class NIX(Coin):
+    NAME = "NIX"
+    SHORTNAME = "NIX"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("26")
+    P2SH_VERBYTES = [bytes.fromhex("35")]
+    GENESIS_HASH = ('dd28ad86def767c3cfc34267a950d871'
+                    'fc7462bc57ea4a929fc3596d9b598e41')
+    DESERIALIZER = lib_tx.DeserializerSegWit
+    TX_COUNT = 114240
+    TX_COUNT_HEIGHT = 87846
+    TX_PER_BLOCK = 3
+    RPC_PORT = 6215
+    REORG_LIMIT = 1000
+
+
+class NIXTestnet(NIX):
+    SHORTNAME = "tNIX"
+    NET = "testnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    GENESIS_HASH = ('dd28ad86def767c3cfc34267a950d871'
+                    'fc7462bc57ea4a929fc3596d9b598e41')
+    P2PKH_VERBYTE = bytes.fromhex("01")
+    P2SH_VERBYTE = [bytes.fromhex("03")]
+    RPC_PORT = 16215
+    DESERIALIZER = lib_tx.DeserializerSegWit
+
+
+class Noir(Coin):
+    NAME = "Noir"
+    SHORTNAME = "NOR"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2SH_VERBYTES = [bytes.fromhex("07")]
+    WIF_BYTE = bytes.fromhex("D0")
+    GENESIS_HASH = ('23911212a525e3d149fcad6c559c8b17'
+                    'f1e8326a272a75ff9bb315c8d96433ef')
+    RPC_PORT = 8825
+    TX_COUNT = 586369
+    TX_COUNT_HEIGHT = 379290
+    TX_PER_BLOCK = 5
+
+
+class BitcoinPlus(Coin):
+    NAME = "BitcoinPlus"
+    SHORTNAME = "XBC"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    P2PKH_VERBYTE = bytes.fromhex("19")
+    P2SH_VERBYTES = [bytes.fromhex("55")]
+    WIF_BYTE = bytes.fromhex("99")
+    GENESIS_HASH = ('0000005f6a28e686f641c616e56182d1'
+                    'b43afbe08a223f23bda23cdf9d55b882')
+    DESERIALIZER = lib_tx.DeserializerTxTime
+    DAEMON = daemon.LegacyRPCDaemon
+    TX_COUNT = 1479247
+    TX_COUNT_HEIGHT = 749740
+    TX_PER_BLOCK = 2
+    RPC_PORT = 8885
+    REORG_LIMIT = 2000
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        import x13_hash
+        return x13_hash.getPoWHash(header)
+
+
+class Myriadcoin(AuxPowMixin, Coin):
+    NAME = "Myriadcoin"
+    SHORTNAME = "XMY"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("32")
+    P2SH_VERBYTES = [bytes.fromhex("09")]
+    WIF_BYTE = bytes.fromhex("b2")
+    GENESIS_HASH = ('00000ffde4c020b5938441a0ea3d314b'
+                    'f619eff0b38f32f78f7583cffa1ea485')
+    DESERIALIZER = lib_tx.DeserializerAuxPowSegWit
+    TX_COUNT = 1976629
+    TX_COUNT_HEIGHT = 2580356
+    TX_PER_BLOCK = 20
+    REORG_LIMIT = 2000
+    RPC_PORT = 10889
+
+
+class MyriadcoinTestnet(Myriadcoin):
+    NAME = "Myriadcoin"
+    SHORTNAME = "XMT"
+    NET = "testnet"
+    XPUB_VERBYTES = bytes.fromhex("043587cf")
+    XPRV_VERBYTES = bytes.fromhex("04358394")
+    P2PKH_VERBYTE = bytes.fromhex("58")
+    P2SH_VERBYTES = [bytes.fromhex("bc")]
+    WIF_BYTE = bytes.fromhex("ef")
+    GENESIS_HASH = ('0000017ce2a79c8bddafbbe47c004aa9'
+                    '2b20678c354b34085f62b762084b9788')
+
+
+class Sparks(Coin):
+    NAME = "Sparks"
+    SHORTNAME = "SPK"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    GENESIS_HASH = ('00000a5c6ddfaac5097218560d5b92d4'
+                    '16931cfeba1abf10c81d1d6a232fc8ea')
+    P2PKH_VERBYTE = bytes.fromhex("26")
+    P2SH_VERBYTES = [bytes.fromhex("0A")]
+    WIF_BYTE = bytes.fromhex("C6")
+    TX_COUNT_HEIGHT = 117400
+    TX_COUNT = 162310
+    TX_PER_BLOCK = 4
+    RPC_PORT = 8818
+    SESSIONCLS = DashElectrumX
+    DAEMON = daemon.DashDaemon
+
+    @classmethod
+    def header_hash(cls, header):
+        import neoscrypt
+        return neoscrypt.getPoWHash(header)
+
+
+# Source: https://github.com/LIMXTEC/BitSend
+class Bitsend(Coin):
+    NAME = "Bitsend"
+    SHORTNAME = "BSD"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    P2PKH_VERBYTE = bytes.fromhex("66")
+    WIF_BYTE = bytes.fromhex("cc")
+    GENESIS_HASH = ('0000012e1b8843ac9ce8c18603658eaf'
+                    '8895f99d3f5e7e1b7b1686f35e3c087a')
+    TX_COUNT = 974672
+    TX_COUNT_HEIGHT = 586022
+    TX_PER_BLOCK = 2
+    RPC_PORT = 8800
+    REORG_LIMIT = 1000
+    DESERIALIZER = lib_tx.DeserializerSegWit
+    XEVAN_TIMESTAMP = 1477958400
+    PEERS = [
+        'ele1.bitsend.cc s t',
+        '51.15.121.233 s t'
+    ]
+
+    @classmethod
+    def header_hash(cls, header):
+        timestamp, = util.unpack_le_uint32_from(header, 68)
+        if timestamp > cls.XEVAN_TIMESTAMP:
+            import xevan_hash
+            return xevan_hash.getPoWHash(header)
+        else:
+            import x11_hash
+            return x11_hash.getPoWHash(header)
+
+    @classmethod
+    def genesis_block(cls, block):
+        header = cls.block_header(block, 0)
+        header_hex_hash = hash_to_hex_str(cls.header_hash(header))
+        if header_hex_hash != cls.GENESIS_HASH:
+            raise CoinError('genesis block has hash {} expected {}'
+                            .format(header_hex_hash, cls.GENESIS_HASH))
+        return header + bytes(1)
+
+
+class Ritocoin(Coin):
+    NAME = "Ritocoin"
+    SHORTNAME = "RITO"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0534E7CA")
+    XPRV_VERBYTES = bytes.fromhex("05347EAC")
+    P2PKH_VERBYTE = bytes.fromhex("19")
+    P2SH_VERBYTES = [bytes.fromhex("69")]
+    GENESIS_HASH = ('00000075e344bdf1c0e433f453764b18'
+                    '30a7aa19b2a5213e707502a22b779c1b')
+    DESERIALIZER = lib_tx.DeserializerSegWit
+    TX_COUNT = 1188090
+    TX_COUNT_HEIGHT = 296030
+    TX_PER_BLOCK = 3
+    RPC_PORT = 8766
+    REORG_LIMIT = 55
+    PEERS = [
+        'electrum-rito.minermore.com s t'
+    ]
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        import x21s_hash
+        return x21s_hash.getPoWHash(header)
+
+
+class Ravencoin(Coin):
+    NAME = "Ravencoin"
+    SHORTNAME = "RVN"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    P2PKH_VERBYTE = bytes.fromhex("3C")
+    P2SH_VERBYTES = [bytes.fromhex("7A")]
+    GENESIS_HASH = ('0000006b444bc2f2ffe627be9d9e7e7a'
+                    '0730000870ef6eb6da46c8eae389df90')
+    DESERIALIZER = lib_tx.DeserializerSegWit
+    X16RV2_ACTIVATION_TIME = 1569945600  # algo switch to x16rv2 at this timestamp
+    TX_COUNT = 5626682
+    TX_COUNT_HEIGHT = 887000
+    TX_PER_BLOCK = 6
+    RPC_PORT = 8766
+    REORG_LIMIT = 55
+    PEERS = [
+    ]
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        timestamp = util.unpack_le_uint32_from(header, 68)[0]
+        if timestamp >= cls.X16RV2_ACTIVATION_TIME:
+            import x16rv2_hash
+            return x16rv2_hash.getPoWHash(header)
+        else:
+            import x16r_hash
+            return x16r_hash.getPoWHash(header)
+
+
+class RavencoinTestnet(Ravencoin):
+    NET = "testnet"
+    XPUB_VERBYTES = bytes.fromhex("043587CF")
+    XPRV_VERBYTES = bytes.fromhex("04358394")
+    P2PKH_VERBYTE = bytes.fromhex("6F")
+    P2SH_VERBYTES = [bytes.fromhex("C4")]
+    WIF_BYTE = bytes.fromhex("EF")
+    GENESIS_HASH = ('000000ecfc5e6324a079542221d00e10'
+                    '362bdc894d56500c414060eea8a3ad5a')
+    X16RV2_ACTIVATION_TIME = 1567533600
+    TX_COUNT = 496158
+    TX_COUNT_HEIGHT = 420500
+    TX_PER_BLOCK = 1
+    RPC_PORT = 18766
+    PEER_DEFAULT_PORTS = {'t': '50003', 's': '50004'}
+    REORG_LIMIT = 55
+    PEERS = [
+    ]
+
+
+class Bolivarcoin(Coin):
+    NAME = "Bolivarcoin"
+    SHORTNAME = "BOLI"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    P2PKH_VERBYTE = bytes.fromhex("55")
+    P2SH_VERBYTES = [bytes.fromhex("05")]
+    WIF_BYTE = bytes.fromhex("D5")
+    GENESIS_HASH = ('00000e4fc293a1912b9d73cbb8d8f727'
+                    '0007a7d84382f1370661e65d5d57b1f6')
+    TX_COUNT = 1082515
+    TX_COUNT_HEIGHT = 540410
+    TX_PER_BLOCK = 10
+    RPC_PORT = 3563
+    REORG_LIMIT = 800
+    PEERS = []
+    SESSIONCLS = DashElectrumX
+    DAEMON = daemon.DashDaemon
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        import x11_hash
+        return x11_hash.getPoWHash(header)
+
+
+class Onixcoin(Coin):
+    NAME = "Onixcoin"
+    SHORTNAME = "ONX"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488B21E")
+    XPRV_VERBYTES = bytes.fromhex("0488ADE4")
+    P2PKH_VERBYTE = bytes.fromhex("4B")
+    GENESIS_HASH = ('000007140b7a6ca0b64965824f5731f6'
+                    'e86daadf19eb299033530b1e61236e43')
+    TX_COUNT = 431808
+    TX_COUNT_HEIGHT = 321132
+    TX_PER_BLOCK = 10
+    RPC_PORT = 41019
+    REORG_LIMIT = 800
+    PEERS = []
+    SESSIONCLS = DashElectrumX
+    DAEMON = daemon.DashDaemon
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        import x11_hash
+        return x11_hash.getPoWHash(header)
+
+
+class Electra(Coin):
+    NAME = "Electra"
+    SHORTNAME = "ECA"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("21")
+    P2SH_VERBYTES = [bytes.fromhex("28")]
+    WIF_BYTE = bytes.fromhex("A1")
+    GENESIS_HASH = ('00000f98da995de0ef1665c7d3338687'
+                    '923c1199230a44ecbdb5cec9306e4f4e')
+    RPC_PORT = 5788
+    TX_COUNT = 615729
+    TX_COUNT_HEIGHT = 205243
+    TX_PER_BLOCK = 3
+    REORG_LIMIT = 100
+    DESERIALIZER = lib_tx.DeserializerElectra
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        version, = util.unpack_le_uint32_from(header)
+        import nist5_hash
+
+        if version != 8:
+            return nist5_hash.getPoWHash(header)
+        else:
+            return double_sha256(header)
+
+
+class ECCoin(Coin):
+    NAME = "ECCoin"
+    SHORTNAME = "ECC"
+    NET = "mainnet"
+    DESERIALIZER = lib_tx.DeserializerECCoin
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("21")
+    P2SH_VERBYTES = [bytes.fromhex("08")]
+    GENESIS_HASH = ('a60ac43c88dbc44b826cf315352a8a7b373d2af8b6e1c4c4a0638859c5e9ecd1')
+    TX_COUNT = 4661197
+    TX_COUNT_HEIGHT = 2114846
+    TX_PER_BLOCK = 10
+    VALUE_PER_COIN = 1000000
+    RPC_PORT = 19119
+
+    @classmethod
+    def header_hash(cls, header):
+        # you have to install scryp python module (pip install scrypt)
+        import scrypt
+        return scrypt.hash(header, header, 1024, 1, 1, 32)
+
+
+class Bellcoin(Coin):
+    NAME = "Bellcoin"
+    SHORTNAME = "BELL"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("19")
+    P2SH_VERBYTES = [bytes.fromhex("55")]
+    WIF_BYTE = bytes.fromhex("80")
+    GENESIS_HASH = ('000008f3b6bd10c2d03b06674a006b8d'
+                    '9731f6cb58179ef1eee008cee2209603')
+    DESERIALIZER = lib_tx.DeserializerSegWit
+    TX_COUNT = 264129
+    TX_COUNT_HEIGHT = 219574
+    TX_PER_BLOCK = 5
+    RPC_PORT = 25252
+    REORG_LIMIT = 1000
+    PEERS = [
+        'bell.electrumx.japanesecoin-pool.work s t',
+        'bell.streetcrypto7.com s t',
+    ]
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        import bell_yespower
+        return bell_yespower.getPoWHash(header)
+
+
+class CPUchain(Coin):
+    NAME = "CPUchain"
+    SHORTNAME = "CPU"
+    NET = "mainnet"
+    P2PKH_VERBYTE = bytes.fromhex("1C")
+    P2SH_VERBYTES = [bytes.fromhex("1E")]
+    GENESIS_HASH = ('000024d8766043ea0e1c9ad42e7ea4b5'
+                    'fdb459887bd80b8f9756f3d87e128f12')
+    DESERIALIZER = lib_tx.DeserializerSegWit
+    TX_COUNT = 4471
+    TX_COUNT_HEIGHT = 3491
+    TX_PER_BLOCK = 2
+    RPC_PORT = 19707
+    REORG_LIMIT = 1000
+    PEERS = [
+        'electrumx.cpuchain.org s t',
+    ]
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        import cpupower
+        return cpupower.getPoWHash(header)
+
+
+class Xaya(NameIndexMixin, AuxPowMixin, Coin):
+    NAME = "Xaya"
+    SHORTNAME = "CHI"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("1c")
+    P2SH_VERBYTES = [bytes.fromhex("1e")]
+    WIF_BYTE = bytes.fromhex("82")
+    GENESIS_HASH = ('e5062d76e5f50c42f493826ac9920b63'
+                    'a8def2626fd70a5cec707ec47a4c4651')
+    TX_COUNT = 1147749
+    TX_COUNT_HEIGHT = 1030000
+    TX_PER_BLOCK = 2
+    DESERIALIZER = lib_tx.DeserializerXaya
+    TRUNCATED_HEADER_SIZE = 80 + 5
+    RPC_PORT = 8396
+    PEERS = [
+        'seeder.xaya.io s50002',
+        'xaya.domob.eu s50002',
+    ]
+
+    # Op-codes for name operations
+    OP_NAME_REGISTER = OpCodes.OP_1
+    OP_NAME_UPDATE = OpCodes.OP_2
+
+    # Valid name prefixes.
+    NAME_REGISTER_OPS = [OP_NAME_REGISTER, "name", -1, OpCodes.OP_2DROP,
+                         OpCodes.OP_DROP]
+    NAME_UPDATE_OPS = [OP_NAME_UPDATE, "name", -1, OpCodes.OP_2DROP,
+                       OpCodes.OP_DROP]
+    NAME_OPERATIONS = [
+        NAME_REGISTER_OPS,
+        NAME_UPDATE_OPS,
+    ]
+
+    @classmethod
+    def genesis_block(cls, block):
+        super().genesis_block(block)
+
+        # In Xaya, the genesis block's coinbase is spendable.  Thus unlike
+        # the generic genesis_block() method, we return the full block here.
+        return block
+
+
+class XayaTestnet(Xaya):
+    SHORTNAME = "XCH"
+    NET = "testnet"
+    P2PKH_VERBYTE = bytes.fromhex("58")
+    P2SH_VERBYTES = [bytes.fromhex("5a")]
+    WIF_BYTE = bytes.fromhex("e6")
+    GENESIS_HASH = ('5195fc01d0e23d70d1f929f21ec55f47'
+                    'e1c6ea1e66fae98ee44cbbc994509bba')
+    TX_COUNT = 51557
+    TX_COUNT_HEIGHT = 49000
+    TX_PER_BLOCK = 1
+    RPC_PORT = 18396
+    PEERS = []
+
+
+class XayaRegtest(XayaTestnet):
+    NET = "regtest"
+    GENESIS_HASH = ('6f750b36d22f1dc3d0a6e483af453010'
+                    '22646dfc3b3ba2187865f5a7d6d83ab1')
+    RPC_PORT = 18493
+
+# Source: https://github.com/GZR0/GRZ0
+
+
+class GravityZeroCoin(ScryptMixin, Coin):
+    NAME = "GravityZeroCoin"
+    SHORTNAME = "GZRO"
+    NET = "mainnet"
+    P2PKH_VERBYTE = bytes.fromhex("26")
+    WIF_BYTE = bytes.fromhex("26")
+    GENESIS_HASH = ('0000028bfbf9ccaed8f28b3ca6b3ffe6b65e29490ab0e4430679bf41cc7c164f')
+    DAEMON = daemon.FakeEstimateLegacyRPCDaemon
+    TX_COUNT = 100
+    TX_COUNT_HEIGHT = 747635
+    TX_PER_BLOCK = 2
+    RPC_PORT = 36442
+    ESTIMATE_FEE = 0.01
+    RELAY_FEE = 0.01
+
+
+class Simplicity(Coin):
+    NAME = "Simplicity"
+    SHORTNAME = "SPL"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0444d5bc")
+    XPRV_VERBYTES = bytes.fromhex("0444f0a3")
+    P2PKH_VERBYTE = bytes.fromhex("12")
+    P2SH_VERBYTE = bytes.fromhex("3b")
+    WIF_BYTE = bytes.fromhex("5d")
+    GENESIS_HASH = ('f4bbfc518aa3622dbeb8d2818a606b82c2b8b1ac2f28553ebdb6fc04d7abaccf')
+    RPC_PORT = 11958
+    TX_COUNT = 1726548
+    TX_COUNT_HEIGHT = 1040000
+    TX_PER_BLOCK = 5
+    REORG_LIMIT = 100
+    DESERIALIZER = lib_tx.DeserializerSimplicity
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        version, = util.unpack_le_uint32_from(header)
+
+        if version < 2:
+            import quark_hash
+            return quark_hash.getPoWHash(header)
+        else:
+            return double_sha256(header)
+
+
+class Myce(Coin):
+    NAME = "Myce"
+    SHORTNAME = "YCE"
+    NET = "mainnet"
+    XPUB_VERBYTES = bytes.fromhex("0488b21e")
+    XPRV_VERBYTES = bytes.fromhex("0488ade4")
+    P2PKH_VERBYTE = bytes.fromhex("32")
+    P2SH_VERBYTE = bytes.fromhex("55")
+    WIF_BYTE = bytes.fromhex("99")
+    GENESIS_HASH = ('0000c74cc66c72cb1a327c5c1d4893ae5276aa50be49fb23cec21df1a2f20d87')
+    RPC_PORT = 23512
+    TX_COUNT = 1568977
+    TX_COUNT_HEIGHT = 774450
+    TX_PER_BLOCK = 3
+    REORG_LIMIT = 100
+    DESERIALIZER = lib_tx.DeserializerSimplicity
+
+    @classmethod
+    def header_hash(cls, header):
+        '''Given a header return the hash.'''
+        version, = util.unpack_le_uint32_from(header)
+
+        if version < 7:
+            import scrypt
+            return scrypt.hash(header, header, 1024, 1, 1, 32)
+        else:
+            return double_sha256(header)
