@@ -1,4 +1,4 @@
-# Copyright (c) 2016, Neil Booth
+# Copyright (c) 2016-2020, Neil Booth
 # Copyright (c) 2017, the ElectrumX authors
 #
 # All rights reserved.
@@ -16,15 +16,18 @@ import time
 from bisect import bisect_right
 from collections import namedtuple
 from glob import glob
-from struct import pack, unpack
+from struct import Struct
 
 import attr
 from aiorpcx import run_in_thread, sleep
 
 import electrumx.lib.util as util
-from electrumx.lib.hash import hash_to_hex_str, HASHX_LEN
+from electrumx.lib.hash import hash_to_hex_str
 from electrumx.lib.merkle import Merkle, MerkleCache
-from electrumx.lib.util import formatted_time
+from electrumx.lib.util import (
+    formatted_time, pack_be_uint16, pack_be_uint32, pack_le_uint64, pack_le_uint32,
+    unpack_le_uint32, unpack_be_uint32, unpack_le_uint64
+)
 from electrumx.server.storage import db_class
 from electrumx.server.history import History
 
@@ -52,7 +55,7 @@ class DB(object):
     it was shutdown uncleanly.
     '''
 
-    DB_VERSIONS = [6]
+    DB_VERSIONS = [6, 7]
 
     class DBError(Exception):
         '''Raised on general DB errors generally indicating corruption.'''
@@ -76,8 +79,18 @@ class DB(object):
         self.db_class = db_class(self.env.db_engine)
         self.history = History()
         self.utxo_db = None
+        self.utxo_flush_count = 0
+        self.fs_height = -1
+        self.fs_tx_count = 0
+        self.db_height = -1
+        self.db_tx_count = 0
+        self.db_tip = None
         self.tx_counts = None
         self.last_flush = time.time()
+        self.last_flush_tx_count = 0
+        self.wall_time = 0
+        self.first_sync = True
+        self.db_version = -1
 
         self.logger.info(f'using {self.env.db_engine} for DB backend')
 
@@ -288,7 +301,7 @@ class DB(object):
         for key, value in flush_data.adds.items():
             # suffix = tx_idx + tx_num
             hashX = value[:-12]
-            suffix = key[-2:] + value[-12:-8]
+            suffix = key[-4:] + value[-12:-8]
             batch_put(b'h' + key[:4] + suffix, hashX)
             batch_put(b'u' + hashX + suffix, value[-8:])
         flush_data.adds.clear()
@@ -348,7 +361,7 @@ class DB(object):
         offsets = []
         for h in headers:
             offset += len(h)
-            offsets.append(pack("<Q", offset))
+            offsets.append(pack_le_uint64(offset))
         # For each header we get the offset of the next header, hence we
         # start writing from the next height
         pos = (height_start + 1) * 8
@@ -356,7 +369,7 @@ class DB(object):
 
     def dynamic_header_offset(self, height):
         assert not self.coin.STATIC_BLOCK_HEADERS
-        offset, = unpack('<Q', self.headers_offsets_file.read(height * 8, 8))
+        offset, = unpack_le_uint64(self.headers_offsets_file.read(height * 8, 8))
         return offset
 
     def dynamic_header_len(self, height):
@@ -402,7 +415,7 @@ class DB(object):
         return await run_in_thread(read_headers)
 
     def fs_tx_hash(self, tx_num):
-        '''Return a par (tx_hash, tx_height) for the given tx number.
+        '''Return a pair (tx_hash, tx_height) for the given tx number.
 
         If the tx_height is not on disk, returns (None, tx_height).'''
         tx_height = bisect_right(self.tx_counts, tx_num)
@@ -411,6 +424,25 @@ class DB(object):
         else:
             tx_hash = self.hashes_file.read(tx_num * 32, 32)
         return tx_hash, tx_height
+
+    def fs_tx_hashes_at_blockheight(self, block_height):
+        '''Return a list of tx_hashes at given block height,
+        in the same order as in the block.
+        '''
+        if block_height > self.db_height:
+            raise self.DBError(f'block {block_height:,d} not on disk (>{self.db_height:,d})')
+        assert block_height >= 0
+        if block_height > 0:
+            first_tx_num = self.tx_counts[block_height - 1]
+        else:
+            first_tx_num = 0
+        num_txs_in_block = self.tx_counts[block_height] - first_tx_num
+        tx_hashes = self.hashes_file.read(first_tx_num * 32, num_txs_in_block * 32)
+        assert num_txs_in_block == len(tx_hashes) // 32
+        return [tx_hashes[idx * 32: (idx+1) * 32] for idx in range(num_txs_in_block)]
+
+    async def tx_hashes_at_blockheight(self, block_height):
+        return await run_in_thread(self.fs_tx_hashes_at_blockheight, block_height)
 
     async def fs_block_hashes(self, height, count):
         headers_concat, headers_count = await self.read_headers(height, count)
@@ -454,7 +486,7 @@ class DB(object):
 
     def undo_key(self, height):
         '''DB key for undo information at the given height.'''
-        return b'U' + pack('>I', height)
+        return b'U' + pack_be_uint32(height)
 
     def read_undo_info(self, height):
         '''Read undo information from a file for the current height.'''
@@ -493,8 +525,8 @@ class DB(object):
         prefix = b'U'
         min_height = self.min_undo_height(self.db_height)
         keys = []
-        for key, hist in self.utxo_db.iterator(prefix=prefix):
-            height, = unpack('>I', key[-4:])
+        for key, _hist in self.utxo_db.iterator(prefix=prefix):
+            height, = unpack_be_uint32(key[-4:])
             if height >= min_height:
                 break
             keys.append(key)
@@ -559,6 +591,10 @@ class DB(object):
         self.fs_tx_count = self.db_tx_count
         self.last_flush_tx_count = self.fs_tx_count
 
+        # Upgrade DB
+        if self.db_version != max(self.DB_VERSIONS):
+            self.upgrade_db()
+
         # Log some stats
         self.logger.info('DB version: {:d}'.format(self.db_version))
         self.logger.info('coin: {}'.format(self.coin.NAME))
@@ -571,6 +607,66 @@ class DB(object):
         if self.first_sync:
             self.logger.info('sync time so far: {}'
                              .format(util.formatted_time(self.wall_time)))
+
+    def upgrade_db(self):
+        self.logger.info('DB version: {:d}'.format(self.db_version))
+        self.logger.info('Upgrading your DB; this can take some time...')
+
+        def upgrade_u_prefix(prefix):
+            count = 0
+            with self.utxo_db.write_batch() as batch:
+                batch_delete = batch.delete
+                batch_put = batch.put
+                # Key: b'u' + address_hashX + tx_idx + tx_num
+                for db_key, db_value in self.utxo_db.iterator(prefix=prefix):
+                    if len(db_key) != 18:
+                        break
+                    count += 1
+                    batch_delete(db_key)
+                    batch_put(db_key[:14] + b'\0\0' + db_key[14:], db_value)
+            return count
+
+        last = time.time()
+        count = 0
+        for cursor in range(65536):
+            prefix = b'u' + pack_be_uint16(cursor)
+            count += upgrade_u_prefix(prefix)
+            now = time.time()
+            if now > last + 10:
+                last = now
+                self.logger.info(f'DB 1 of 2: {count:,d} entries updated, '
+                                 f'{cursor * 100 / 65536:.1f}% complete')
+        self.logger.info('DB 1 of 2 upgraded successfully')
+
+        def upgrade_h_prefix(prefix):
+            count = 0
+            with self.utxo_db.write_batch() as batch:
+                batch_delete = batch.delete
+                batch_put = batch.put
+                # Key: b'h' + compressed_tx_hash + tx_idx + tx_num
+                for db_key, db_value in self.utxo_db.iterator(prefix=prefix):
+                    if len(db_key) != 11:
+                        break
+                    count += 1
+                    batch_delete(db_key)
+                    batch_put(db_key[:7] + b'\0\0' + db_key[7:], db_value)
+            return count
+
+        last = time.time()
+        count = 0
+        for cursor in range(65536):
+            prefix = b'h' + pack_be_uint16(cursor)
+            count += upgrade_h_prefix(prefix)
+            now = time.time()
+            if now > last + 10:
+                last = now
+                self.logger.info(f'DB 2 of 2: {count:,d} entries updated, '
+                                 f'{cursor * 100 / 65536:.1f}% complete')
+
+        self.db_version = max(self.DB_VERSIONS)
+        with self.utxo_db.write_batch() as batch:
+            self.write_utxo_state(batch)
+        self.logger.info('DB 2 of 2 upgraded successfully')
 
     def write_utxo_state(self, batch):
         '''Write (UTXO) state to the batch.'''
@@ -596,13 +692,13 @@ class DB(object):
         def read_utxos():
             utxos = []
             utxos_append = utxos.append
-            s_unpack = unpack
+            unpack_2_le_uint32 = Struct('<II').unpack
             # Key: b'u' + address_hashX + tx_idx + tx_num
             # Value: the UTXO value as a 64-bit unsigned integer
             prefix = b'u' + hashX
             for db_key, db_value in self.utxo_db.iterator(prefix=prefix):
-                tx_pos, tx_num = s_unpack('<HI', db_key[-6:])
-                value, = unpack('<Q', db_value)
+                tx_pos, tx_num = unpack_2_le_uint32(db_key[-8:])
+                value, = unpack_le_uint64(db_value)
                 tx_hash, height = self.fs_tx_hash(tx_num)
                 utxos_append(UTXO(tx_num, tx_pos, tx_hash, height, value))
             return utxos
@@ -626,7 +722,7 @@ class DB(object):
             for each prevout.
             '''
             def lookup_hashX(tx_hash, tx_idx):
-                idx_packed = pack('<H', tx_idx)
+                idx_packed = pack_le_uint32(tx_idx)
 
                 # Key: b'h' + compressed_tx_hash + tx_idx + tx_num
                 # Value: hashX
@@ -635,8 +731,8 @@ class DB(object):
                 # Find which entry, if any, the TX_HASH matches.
                 for db_key, hashX in self.utxo_db.iterator(prefix=prefix):
                     tx_num_packed = db_key[-4:]
-                    tx_num, = unpack('<I', tx_num_packed)
-                    hash, height = self.fs_tx_hash(tx_num)
+                    tx_num, = unpack_le_uint32(tx_num_packed)
+                    hash, _height = self.fs_tx_hash(tx_num)
                     if hash == tx_hash:
                         return hashX, idx_packed + tx_num_packed
                 return None, None
@@ -657,7 +753,7 @@ class DB(object):
                     # This can happen if the DB was updated between
                     # getting the hashXs and getting the UTXOs
                     return None
-                value, = unpack('<Q', db_value)
+                value, = unpack_le_uint64(db_value)
                 return hashX, value
             return [lookup_utxo(*hashX_pair) for hashX_pair in hashX_pairs]
 
