@@ -63,7 +63,7 @@ class API(MemPoolAPI):
 
     def __init__(self):
         self._height = 0
-        self._cached_height = self._height
+        self._cached_height = self._db_height = self._height
         # Create a pool of hash160s.  Map them to their script hashes
         # Create a bunch of UTXOs paying to those script hashes
         # Create a bunch of TXs that spend from the UTXO set and create
@@ -203,6 +203,9 @@ class API(MemPoolAPI):
         self._cached_height = self._height
         return self._height
 
+    def db_height(self):
+        return self._db_height
+
     def cached_height(self):
         return self._cached_height
 
@@ -272,7 +275,8 @@ async def test_keep_synchronized(caplog):
     assert in_caplog(caplog, 'beginning processing of daemon mempool')
     assert in_caplog(caplog, 'compact fee histogram')
     assert in_caplog(caplog, 'synced in ')
-    assert in_caplog(caplog, '0 txs touching 0 addresses')
+    assert in_caplog(caplog, '0 txs')
+    assert in_caplog(caplog, 'touching 0 addresses')
     assert not in_caplog(caplog, 'txs dropped')
 
 
@@ -443,7 +447,7 @@ async def test_daemon_drops_txs():
 
 
 @pytest.mark.asyncio
-async def test_notifications():
+async def test_notifications(caplog):
     # Tests notifications over a cycle of:
     # 1) A first batch of txs come in
     # 2) A second batch of txs come in
@@ -461,6 +465,8 @@ async def test_notifications():
     second_hashes = api.ordered_adds[n:]
     second_touched = api.touched(second_hashes)
 
+    caplog.set_level(logging.DEBUG)
+
     async with TaskGroup() as group:
         # First batch enters the mempool
         api.raw_txs = {hash: raw_txs[hash] for hash in first_hashes}
@@ -471,7 +477,7 @@ async def test_notifications():
         await event.wait()
         assert len(api.on_mempool_calls) == 1
         touched, height = api.on_mempool_calls[0]
-        assert height == api._height == api._cached_height
+        assert height == api._height == api._db_height == api._cached_height
         assert touched == first_touched
         # Second batch enters the mempool
         api.raw_txs = raw_txs
@@ -479,21 +485,32 @@ async def test_notifications():
         await event.wait()
         assert len(api.on_mempool_calls) == 2
         touched, height = api.on_mempool_calls[1]
-        assert height == api._height == api._cached_height
+        assert height == api._height == api._db_height == api._cached_height
         # Touched is incremental
         assert touched == second_touched
         # Block found; first half confirm
         new_height = 2
         api._height = new_height
-        api.db_utxos.update(first_utxos)
-        for spend in first_spends:
-            del api.db_utxos[spend]
         api.raw_txs = {hash: raw_txs[hash] for hash in second_hashes}
         api.txs = {hash: txs[hash] for hash in second_hashes}
+        # Delay the DB update
+        assert not in_caplog(caplog, 'waiting for DB to sync')
+        async with ignore_after(mempool.refresh_secs * 2):
+            await event.wait()
+        assert in_caplog(caplog, 'waiting for DB to sync')
+        assert len(api.on_mempool_calls) == 2
+        assert not event.is_set()
+        assert api._height == api._cached_height == new_height
+        assert touched == second_touched
+        # Now update the DB
+        api.db_utxos.update(first_utxos)
+        api._db_height = new_height
+        for spend in first_spends:
+            del api.db_utxos[spend]
         await event.wait()
         assert len(api.on_mempool_calls) == 3
         touched, height = api.on_mempool_calls[2]
-        assert height == api._height == api._cached_height == new_height
+        assert height == api._db_height == new_height
         assert touched == first_touched
         await group.cancel_remaining()
 
